moves test data into separate files
