+++ test/data/base-runners.js
@@ -1,0 +1,5 @@
+module.exports = {
+  runnerFirst: 'hit', // from hits
+  runners1and2: 'stealHome', // from steals
+};
+
