+++ test/data/game-status.js
@@ -1,0 +1,332 @@
+const beforeFirstPitch = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 2,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 0,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Play ball!",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "",
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": -1,
+  "homeTeamBatterCount": -1
+};
+
+const secondHalf = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 3,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 0,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Bottom of 1, New York Millennials batting.",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "",
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": 2,
+  "homeTeamBatterCount": -1
+};
+
+const firstHalf = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 3,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Top of 2, Hades Tigers batting.",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "",
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": 2,
+  "homeTeamBatterCount": 3
+};
+
+const endOld = {
+  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
+  "day": 119,
+  "phase": 6,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 8,
+  "season": 10,
+  "weather": 14,
+  "awayOdds": 0.38125626427142956,
+  "awayOuts": 3,
+  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
+  "homeOdds": 0.6187437357285706,
+  "homeOuts": 3,
+  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
+  "outcomes": [],
+  "awayBalls": 3,
+  "awayBases": 4,
+  "awayScore": 9,
+  "finalized": true,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 5,
+  "playCount": 306,
+  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Game over.",
+  "tournament": -1,
+  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
+  "awayStrikes": 4,
+  "baseRunners": [],
+  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "scoreLedger": "",
+  "scoreUpdate": "",
+  "seriesIndex": 5,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hellmouth Sunbeams",
+  "gameComplete": true,
+  "homeTeamName": "Seattle Garages",
+  "isPostseason": true,
+  "seriesLength": 3,
+  "awayBatterMod": "",
+  "awayTeamColor": "#fffbab",
+  "awayTeamEmoji": "0x1F31E",
+  "basesOccupied": [],
+  "homeBatterMod": "",
+  "homeTeamColor": "#2b4075",
+  "homeTeamEmoji": "0x1F3B8",
+  "awayBatterName": "",
+  "awayPitcherMod": "",
+  "baseRunnerMods": [],
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "homePitcherMod": "",
+  "awayPitcherName": "Lars Taylor",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Arturo Huerta",
+  "awayTeamNickname": "Sunbeams",
+  "homeTeamNickname": "Garages",
+  "awayTeamBatterCount": 39,
+  "homeTeamBatterCount": 36,
+  "awayTeamSecondaryColor": "#fffbab",
+  "homeTeamSecondaryColor": "#5a83ea"
+};
+
+const endNew = {
+  "id": "a1ad2005-7bb6-4f7f-95bf-8260d43b923e",
+  "day": 115,
+  "phase": 7,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 8,
+  "season": 13,
+  "weather": 18,
+  "awayOdds": 0.3979624149734539,
+  "awayOuts": 3,
+  "awayTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
+  "homeOdds": 0.6020375850265461,
+  "homeOuts": 3,
+  "homeTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
+  "outcomes": [],
+  "awayBalls": 4,
+  "awayBases": 4,
+  "awayScore": 8,
+  "finalized": true,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 5,
+  "playCount": 317,
+  "stadiumId": "36b94380-39a6-4e32-b525-3e888215798a",
+  "statsheet": "76d8535d-e9fc-48ee-9674-6c91dccc0184",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Moist Talkers 8, Tacos 5",
+  "tournament": -1,
+  "awayPitcher": "60026a9d-fc9a-4f5a-94fd-2225398fa3da",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "0bb35615-63f2-4492-80ec-b6b322dc5450",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "scoreLedger": "",
+  "scoreUpdate": "",
+  "seriesIndex": 2,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "Canada Moist Talkers",
+  "gameComplete": true,
+  "homeTeamName": "LA Unlimited Tacos",
+  "isPostseason": true,
+  "isTitleMatch": false,
+  "seriesLength": 3,
+  "awayBatterMod": "",
+  "awayTeamColor": "#f5feff",
+  "awayTeamEmoji": "0x1F5E3",
+  "basesOccupied": [],
+  "homeBatterMod": "",
+  "homeTeamColor": "#64376e",
+  "homeTeamEmoji": "0x1F32E",
+  "awayBatterName": "",
+  "awayPitcherMod": "",
+  "baseRunnerMods": [],
+  "gameStartPhase": 19,
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "homePitcherMod": "",
+  "newInningPhase": -1,
+  "topInningScore": 0,
+  "awayPitcherName": "Bright Zimmerman",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Sexton Wheerer",
+  "awayTeamNickname": "Moist Talkers",
+  "homeTeamNickname": "Tacos",
+  "secretBaserunner": null,
+  "bottomInningScore": 0,
+  "awayTeamBatterCount": 41,
+  "homeTeamBatterCount": 41,
+  "awayTeamSecondaryColor": "#f5feff",
+  "homeTeamSecondaryColor": "#b063c1"
+};
+
+
+module.exports = {
+  beforeFirstPitch,
+  secondHalf,
+  firstHalf,
+  endOld,
+  endNew,
+};
+
