+++ test/data/gameEvents.js
@@ -1,3222 +1,33 @@
+const GAME_DATA_SOURCES = [
+  require('./game-status'),
+  require('./hits'),
+  require('./outs'),
+  require('./special'),
+  require('./steals'),
+  require('./walks'),
+  // the following use game event data declared above so need to be processed
+  // last
+  require('./base-runners'),
+  require('./misc'),
+];
+
+const GAME_DATA = {};
+
+for (let source of GAME_DATA_SOURCES) {
+  for (let sourceData in source) {
+    if (typeof(source[sourceData]) === 'string') {
+      GAME_DATA[sourceData] = GAME_DATA[source[sourceData]];
+    } else {
+      GAME_DATA[sourceData] = source[sourceData];
+    }
+  }
+}
+
 const noData = {
-};
-
-const beforeFirstPitch = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 2,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Play ball!",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "",
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": -1,
-  "homeTeamBatterCount": -1
-};
-
-const flyout = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Fish Summer hit a flyout to Penelope Mathews.",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "",
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": 0,
-  "homeTeamBatterCount": -1
-};
-
-const groundOut = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 2,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Spears Taylor hit a ground out to Thomas Dracaena.",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "",
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": 2,
-  "homeTeamBatterCount": -1
-};
-
-const strikeOut = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Mclaughlin Scorpler strikes out looking.",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "",
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": 3,
-  "homeTeamBatterCount": 3
-};
-
-const secondHalf = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Bottom of 1, New York Millennials batting.",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "",
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": 2,
-  "homeTeamBatterCount": -1
-};
-
-const firstHalf = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Top of 2, Hades Tigers batting.",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "",
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": 2,
-  "homeTeamBatterCount": 3
-};
-
-const hit = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Conrad Vaughan hits a Single!",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "5dbf11c0-994a-4482-bd1e-99379148ee45"
-  ],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [
-      0
-  ],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "",
-  "halfInningOuts": 2,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [
-      "Conrad Vaughan"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 0,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": 2,
-  "homeTeamBatterCount": 2
-};
-
-const dbl = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 1,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Ren Morin hits a Double! 1 scores.",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "32c9bce6-6e52-40fa-9f64-3629b3d026a8"
-  ],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [
-      1
-  ],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "",
-  "halfInningOuts": 2,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [
-      "Ren Morin"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 1,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": 6,
-  "homeTeamBatterCount": 3
-};
-
-const homeRun = {
-  "id": "b2cafd20-a799-48f6-abd7-c99bd79a6bd1",
-  "day": 116,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 13,
-  "weather": 18,
-  "awayOdds": 0.44891216386293564,
-  "awayOuts": 3,
-  "awayTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
-  "homeOdds": 0.5510878361370642,
-  "homeOuts": 3,
-  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
-  "outcomes": [],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 3,
-  "playCount": 65,
-  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
-  "statsheet": "a1d470e8-8f19-4186-a64b-73f0195a896f",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Alston Cerveza hits a 2-run home run!\nThe ball lands in a Big Bucket. An extra Run scores!",
-  "tournament": -1,
-  "awayPitcher": "d81ce662-07b6-4a73-baa4-acbbb41f9dc5",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "3 Runs scored!",
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "LA Unlimited Tacos",
-  "gameComplete": false,
-  "homeTeamName": "Canada Moist Talkers",
-  "isPostseason": true,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#64376e",
-  "awayTeamEmoji": "0x1F32E",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#f5feff",
-  "homeTeamEmoji": "0x1F5E3",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 12,
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Yummy Elliott",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 3,
-  "homePitcherName": "Greer Lott",
-  "awayTeamNickname": "Tacos",
-  "homeTeamNickname": "Moist Talkers",
-  "secretBaserunner": null,
-  "bottomInningScore": 3,
-  "awayTeamBatterCount": 8,
-  "homeTeamBatterCount": 6,
-  "awayTeamSecondaryColor": "#b063c1",
-  "homeTeamSecondaryColor": "#f5feff"
-};
-
-const steal = {
-  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
-  "day": 119,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 10,
-  "weather": 14,
-  "awayOdds": 0.38125626427142956,
-  "awayOuts": 3,
-  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
-  "homeOdds": 0.6187437357285706,
-  "homeOuts": 3,
-  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "outcomes": [],
-  "awayBalls": 3,
-  "awayBases": 4,
-  "awayScore": 1,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 43,
-  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
-  "atBatBalls": 0,
-  "awayBatter": "190a0f31-d686-4ac4-a7f3-cfc87b72c145",
-  "homeBatter": null,
-  "lastUpdate": "Sutton Bishop steals fourth base!",
-  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
-  "awayStrikes": 4,
-  "baseRunners": [
-      "4f69e8c2-b2a1-4e98-996a-ccf35ac844c5"
-  ],
-  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "seriesIndex": 5,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hellmouth Sunbeams",
-  "gameComplete": false,
-  "homeTeamName": "Seattle Garages",
-  "isPostseason": true,
-  "seriesLength": 3,
-  "awayTeamColor": "#fffbab",
-  "awayTeamEmoji": "0x1F31E",
-  "basesOccupied": [
-      0
-  ],
-  "homeTeamColor": "#2b4075",
-  "homeTeamEmoji": "0x1F3B8",
-  "awayBatterName": "Nerd Pacheco",
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "awayPitcherName": "Lars Taylor",
-  "baseRunnerNames": [
-      "Igneus Delacruz"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 1,
-  "homePitcherName": "Arturo Huerta",
-  "awayTeamNickname": "Sunbeams",
-  "homeTeamNickname": "Garages",
-  "awayTeamBatterCount": 6,
-  "homeTeamBatterCount": 3,
-  "awayTeamSecondaryColor": "#fffbab",
-  "homeTeamSecondaryColor": "#5a83ea"
-};
-
-const soloHR = {
-  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
-  "day": 119,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 10,
-  "weather": 14,
-  "awayOdds": 0.38125626427142956,
-  "awayOuts": 3,
-  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
-  "homeOdds": 0.6187437357285706,
-  "homeOuts": 3,
-  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "outcomes": [],
-  "awayBalls": 3,
-  "awayBases": 4,
-  "awayScore": 4,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 48,
-  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Hahn Fox hits a solo home run!",
-  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
-  "awayStrikes": 4,
-  "baseRunners": [],
-  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "seriesIndex": 5,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hellmouth Sunbeams",
-  "gameComplete": false,
-  "homeTeamName": "Seattle Garages",
-  "isPostseason": true,
-  "seriesLength": 3,
-  "awayTeamColor": "#fffbab",
-  "awayTeamEmoji": "0x1F31E",
-  "basesOccupied": [],
-  "homeTeamColor": "#2b4075",
-  "homeTeamEmoji": "0x1F3B8",
-  "awayBatterName": "",
-  "halfInningOuts": 2,
-  "homeBatterName": "",
-  "awayPitcherName": "Lars Taylor",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 4,
-  "homePitcherName": "Arturo Huerta",
-  "awayTeamNickname": "Sunbeams",
-  "homeTeamNickname": "Garages",
-  "awayTeamBatterCount": 8,
-  "homeTeamBatterCount": 3,
-  "awayTeamSecondaryColor": "#fffbab",
-  "homeTeamSecondaryColor": "#5a83ea"
-};
-
-const twoRunHR = {
-  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
-  "day": 119,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 10,
-  "weather": 14,
-  "awayOdds": 0.38125626427142956,
-  "awayOuts": 3,
-  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
-  "homeOdds": 0.6187437357285706,
-  "homeOuts": 3,
-  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "outcomes": [],
-  "awayBalls": 3,
-  "awayBases": 4,
-  "awayScore": 4,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 2,
-  "playCount": 64,
-  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Oliver Mueller hits a 2-run home run!",
-  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
-  "awayStrikes": 4,
-  "baseRunners": [],
-  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "seriesIndex": 5,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hellmouth Sunbeams",
-  "gameComplete": false,
-  "homeTeamName": "Seattle Garages",
-  "isPostseason": true,
-  "seriesLength": 3,
-  "awayTeamColor": "#fffbab",
-  "awayTeamEmoji": "0x1F31E",
-  "basesOccupied": [],
-  "homeTeamColor": "#2b4075",
-  "homeTeamEmoji": "0x1F3B8",
-  "awayBatterName": "",
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "awayPitcherName": "Lars Taylor",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 2,
-  "homePitcherName": "Arturo Huerta",
-  "awayTeamNickname": "Sunbeams",
-  "homeTeamNickname": "Garages",
-  "awayTeamBatterCount": 9,
-  "homeTeamBatterCount": 5,
-  "awayTeamSecondaryColor": "#fffbab",
-  "homeTeamSecondaryColor": "#5a83ea"
-};
-
-const triple = {
-  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
-  "day": 119,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 10,
-  "weather": 14,
-  "awayOdds": 0.38125626427142956,
-  "awayOuts": 3,
-  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
-  "homeOdds": 0.6187437357285706,
-  "homeOuts": 3,
-  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "outcomes": [],
-  "awayBalls": 3,
-  "awayBases": 4,
-  "awayScore": 4,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 60,
-  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Summers Pony hits a Triple!",
-  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
-  "awayStrikes": 4,
-  "baseRunners": [
-      "81a0889a-4606-4f49-b419-866b57331383"
-  ],
-  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "seriesIndex": 5,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hellmouth Sunbeams",
-  "gameComplete": false,
-  "homeTeamName": "Seattle Garages",
-  "isPostseason": true,
-  "seriesLength": 3,
-  "awayTeamColor": "#fffbab",
-  "awayTeamEmoji": "0x1F31E",
-  "basesOccupied": [
-      2
-  ],
-  "homeTeamColor": "#2b4075",
-  "homeTeamEmoji": "0x1F3B8",
-  "awayBatterName": "",
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "awayPitcherName": "Lars Taylor",
-  "baseRunnerNames": [
-      "Summers Pony"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 0,
-  "homePitcherName": "Arturo Huerta",
-  "awayTeamNickname": "Sunbeams",
-  "homeTeamNickname": "Garages",
-  "awayTeamBatterCount": 9,
-  "homeTeamBatterCount": 4,
-  "awayTeamSecondaryColor": "#fffbab",
-  "homeTeamSecondaryColor": "#5a83ea"
-};
-
-const hitButNoScore = {
-  "id": "b2cafd20-a799-48f6-abd7-c99bd79a6bd1",
-  "day": 116,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 13,
-  "weather": 18,
-  "awayOdds": 0.44891216386293564,
-  "awayOuts": 3,
-  "awayTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
-  "homeOdds": 0.5510878361370642,
-  "homeOuts": 3,
-  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
-  "outcomes": [],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 3,
-  "playCount": 73,
-  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
-  "statsheet": "a1d470e8-8f19-4186-a64b-73f0195a896f",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Lachlan Shelton hits a Single!",
-  "tournament": -1,
-  "awayPitcher": "d81ce662-07b6-4a73-baa4-acbbb41f9dc5",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "9f85676a-7411-444a-8ae2-c7f8f73c285c"
-  ],
-  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "LA Unlimited Tacos",
-  "gameComplete": false,
-  "homeTeamName": "Canada Moist Talkers",
-  "isPostseason": true,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#64376e",
-  "awayTeamEmoji": "0x1F32E",
-  "basesOccupied": [
-      0
-  ],
-  "homeBatterMod": "",
-  "homeTeamColor": "#f5feff",
-  "homeTeamEmoji": "0x1F5E3",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [
-      ""
-  ],
-  "gameStartPhase": 12,
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Yummy Elliott",
-  "baseRunnerNames": [
-      "Lachlan Shelton"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 3,
-  "homePitcherName": "Greer Lott",
-  "awayTeamNickname": "Tacos",
-  "homeTeamNickname": "Moist Talkers",
-  "secretBaserunner": null,
-  "bottomInningScore": 3,
-  "awayTeamBatterCount": 8,
-  "homeTeamBatterCount": 7,
-  "awayTeamSecondaryColor": "#b063c1",
-  "homeTeamSecondaryColor": "#f5feff"
-};
-
-const stealScore = {
-  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
-  "day": 119,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 10,
-  "weather": 14,
-  "awayOdds": 0.38125626427142956,
-  "awayOuts": 3,
-  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
-  "homeOdds": 0.6187437357285706,
-  "homeOuts": 3,
-  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "outcomes": [],
-  "awayBalls": 3,
-  "awayBases": 4,
-  "awayScore": 1,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 43,
-  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
-  "atBatBalls": 0,
-  "awayBatter": "190a0f31-d686-4ac4-a7f3-cfc87b72c145",
-  "homeBatter": null,
-  "lastUpdate": "Sutton Bishop steals fourth base!",
-  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
-  "awayStrikes": 4,
-  "baseRunners": [
-      "4f69e8c2-b2a1-4e98-996a-ccf35ac844c5"
-  ],
-  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "seriesIndex": 5,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hellmouth Sunbeams",
-  "gameComplete": false,
-  "homeTeamName": "Seattle Garages",
-  "isPostseason": true,
-  "seriesLength": 3,
-  "awayTeamColor": "#fffbab",
-  "awayTeamEmoji": "0x1F31E",
-  "basesOccupied": [
-      0
-  ],
-  "homeTeamColor": "#2b4075",
-  "homeTeamEmoji": "0x1F3B8",
-  "awayBatterName": "Nerd Pacheco",
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "awayPitcherName": "Lars Taylor",
-  "baseRunnerNames": [
-      "Igneus Delacruz"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 1,
-  "homePitcherName": "Arturo Huerta",
-  "awayTeamNickname": "Sunbeams",
-  "homeTeamNickname": "Garages",
-  "awayTeamBatterCount": 6,
-  "homeTeamBatterCount": 3,
-  "awayTeamSecondaryColor": "#fffbab",
-  "homeTeamSecondaryColor": "#5a83ea"
-};
-
-const stealSecond = {
-  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
-  "day": 111,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 3,
-  "weather": 12,
-  "awayOdds": 0.5588613238139439,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.441138676186056,
-  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "outcomes": [],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
-  "atBatBalls": 0,
-  "awayBatter": "32c9bce6-6e52-40fa-9f64-3629b3d026a8",
-  "homeBatter": null,
-  "lastUpdate": "Moody Cookbook steals second base!",
-  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "70ccff1e-6b53-40e2-8844-0a28621cb33e"
-  ],
-  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "New York Millennials",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [
-      1
-  ],
-  "homeTeamColor": "#ffd4d8",
-  "homeTeamEmoji": "0x1F4F1",
-  "awayBatterName": "Ren Morin",
-  "halfInningOuts": 2,
-  "homeBatterName": "",
-  "awayPitcherName": "Yazmin Mason",
-  "baseRunnerNames": [
-      "Moody Cookbook"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 0,
-  "homePitcherName": "Patty Fox",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Millennials",
-  "awayTeamBatterCount": 6,
-  "homeTeamBatterCount": 3
-};
-
-const walk = {
-  "_id": "97d88b9e-406d-4f31-a18f-2a3b903edc03",
-  "day": 110,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 1,
-  "weather": 7,
-  "awayOdds": 0.6557171270683066,
-  "awayTeam": "23e4cbc1-e9cd-47fa-a35b-bfa06f726cb7",
-  "homeOdds": 0.34428287293169346,
-  "homeTeam": "bfd38797-8404-4b38-8b82-341da28b1f83",
-  "outcomes": [],
-  "awayScore": 1,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 0,
-  "statsheet": "9a42badd-8646-4e90-9d33-3e5accc3403f",
-  "atBatBalls": 0,
-  "awayBatter": "",
-  "homeBatter": "248ccf3d-d5f6-4b69-83d9-40230ca909cd",
-  "lastUpdate": "Jessica Telephone draws a walk.",
-  "awayPitcher": "1732e623-ffc2-40f0-87ba-fdcf97131f1f",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "083d09d4-7ed3-4100-b021-8fbe30dd43e8"
-  ],
-  "homePitcher": "2ae8cbfc-2155-4647-9996-3f2591091baf",
-  "homeStrikes": 3,
-  "seriesIndex": 4,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Philly Pies",
-  "gameComplete": false,
-  "homeTeamName": "Charleston Shoe Thieves",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#399d8f",
-  "awayTeamEmoji": "0x1F967",
-  "basesOccupied": [
-      0
-  ],
-  "homeTeamColor": "#ffce0a",
-  "homeTeamEmoji": "0x1F45F",
-  "awayBatterName": "",
-  "halfInningOuts": 1,
-  "homeBatterName": "Antonio Wallace",
-  "awayPitcherName": "Betsy Trombone",
-  "baserunnerCount": 1,
-  "halfInningScore": 0,
-  "homePitcherName": "Forrest Bookbaby",
-  "awayTeamNickname": "Pies",
-  "homeTeamNickname": "Shoe Thieves",
-  "awayTeamBatterCount": 6,
-  "homeTeamBatterCount": 3
-};
-
-const walkScores = {
-  id: 'b6f5af1a-79e8-4f05-a6c7-9b55c42124ab',
-  day: 95,
-  phase: 6,
-  rules: '4ae9d46a-5408-460a-84fb-cbd8d03fff6c',
-  shame: false,
-  inning: 0,
-  season: 13,
-  weather: 13,
-  awayOdds: 0.5538712361406611,
-  awayOuts: 3,
-  awayTeam: 'eb67ae5e-c4bf-46ca-bbbc-425cd34182ff',
-  homeOdds: 0.4461287638593388,
-  homeOuts: 3,
-  homeTeam: 'adc5b394-8f76-416d-9ce9-813706877b84',
-  outcomes: [ 'Rodriguez Internet returned from Elsewhere!' ],
-  awayBalls: 4,
-  awayBases: 4,
-  awayScore: 1,
-  finalized: false,
-  gameStart: true,
-  homeBalls: 4,
-  homeBases: 4,
-  homeScore: 6,
-  playCount: 83,
-  stadiumId: '21ce1233-b2d6-4bd2-8d03-68a8915afdc5',
-  statsheet: '8778af73-a34a-4f67-870f-9701f3e1d0d8',
-  atBatBalls: 0,
-  awayBatter: null,
-  homeBatter: null,
-  lastUpdate: 'Hewitt Best draws a walk.\n' +
-    'Marquez Clark scores!\n' +
-    'Hewitt Best used their Free Refill.\n' +
-    'Hewitt Best Refills the In!',
-  tournament: -1,
-  awayPitcher: '60026a9d-fc9a-4f5a-94fd-2225398fa3da',
-  awayStrikes: 3,
-  baseRunners: [
-    '53e701c7-e3c8-4e18-ba05-9b41b4b64cda',
-    'd6c69d2d-9344-4b19-85a4-6cfcbaead5d2',
-    '88cd6efa-dbf2-4309-aabe-ec1d6f21f98a'
-  ],
-  homePitcher: '138fccc3-e66f-4b07-8327-d4b6f372f654',
-  homeStrikes: 3,
-  repeatCount: 0,
-  scoreLedger: '',
-  scoreUpdate: '1 Run scored!',
-  seriesIndex: 3,
-  terminology: 'b67e9bbb-1495-4e1b-b517-f1444b0a6c8b',
-  topOfInning: false,
-  atBatStrikes: 0,
-  awayTeamName: 'Canada Moist Talkers',
-  gameComplete: false,
-  homeTeamName: 'Kansas City Breath Mints',
-  isPostseason: false,
-  isTitleMatch: false,
-  seriesLength: 3,
-  awayBatterMod: '',
-  awayTeamColor: '#f5feff',
-  awayTeamEmoji: '0x1F5E3',
-  basesOccupied: [ 2, 1, 0 ],
-  homeBatterMod: '',
-  homeTeamColor: '#178f55',
-  homeTeamEmoji: '0x1F36C',
-  awayBatterName: '',
-  awayPitcherMod: '',
-  baseRunnerMods: [ '', '', '' ],
-  gameStartPhase: 11,
-  halfInningOuts: 1,
-  homeBatterName: '',
-  homePitcherMod: '',
-  newInningPhase: -1,
-  topInningScore: 1,
-  awayPitcherName: 'Bright Zimmerman',
-  baseRunnerNames: [ 'Marquez Clark', 'Joshua Watson', 'Hewitt Best' ],
-  baserunnerCount: 3,
-  halfInningScore: 6,
-  homePitcherName: 'Oscar Vaughan',
-  awayTeamNickname: 'Moist Talkers',
-  homeTeamNickname: 'Breath Mints',
-  secretBaserunner: null,
-  bottomInningScore: 6,
-  awayTeamBatterCount: 4,
-  homeTeamBatterCount: 8,
-  awayTeamSecondaryColor: '#f5feff',
-  homeTeamSecondaryColor: '#178f55'
-};
-
-const stealHome = {
-  "_id": "d8f95c60-a1c6-40f6-9d9f-25511a0f2c68",
-  "day": 102,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 8,
-  "season": 1,
-  "weather": 7,
-  "awayOdds": 0.5413947119936758,
-  "awayTeam": "bfd38797-8404-4b38-8b82-341da28b1f83",
-  "homeOdds": 0.4586052880063242,
-  "homeTeam": "b024e975-1c4a-4575-8936-a3754a08806a",
-  "outcomes": [],
-  "awayScore": 11,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 5,
-  "statsheet": "f69a6db4-544a-4d26-ba1e-13fd98c7249a",
-  "atBatBalls": 2,
-  "awayBatter": "18798b8f-6391-4cb2-8a5f-6fb540d646d5",
-  "homeBatter": "f73009c5-2ede-4dc4-b96d-84ba93c8a429",
-  "lastUpdate": "Marco Stink steals home!",
-  "awayPitcher": "36786f44-9066-4028-98d9-4fa84465ab9e",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "17397256-c28c-4cad-85f2-a21768c66e67",
-      "c83f0fe0-44d1-4342-81e8-944bb38f8e23"
-  ],
-  "homePitcher": "740d5fef-d59f-4dac-9a75-739ec07f91cf",
-  "homeStrikes": 3,
-  "seriesIndex": 4,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 1,
-  "awayTeamName": "Charleston Shoe Thieves",
-  "gameComplete": false,
-  "homeTeamName": "Dallas Steaks",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#ffce0a",
-  "awayTeamEmoji": "0x1F45F",
-  "basesOccupied": [
-      1,
-      0
-  ],
-  "homeTeamColor": "#8c8d8f",
-  "homeTeamEmoji": "0x1F969",
-  "awayBatterName": "Morrow Doyle",
-  "halfInningOuts": 2,
-  "homeBatterName": "Thomas Kirby",
-  "awayPitcherName": "Beasley Gloom",
-  "baserunnerCount": 2,
-  "halfInningScore": 1,
-  "homePitcherName": "Conner Haley",
-  "awayTeamNickname": "Shoe Thieves",
-  "homeTeamNickname": "Steaks",
-  "awayTeamBatterCount": 45,
-  "homeTeamBatterCount": 38
-};
-
-const caughtStealing = {
-  "_id": "72418a6e-b2c0-44dc-bc9f-e045e049411b",
-  "day": 104,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 6,
-  "season": 1,
-  "weather": 7,
-  "awayOdds": 0.4803003352967402,
-  "awayTeam": "23e4cbc1-e9cd-47fa-a35b-bfa06f726cb7",
-  "homeOdds": 0.5196996647032597,
-  "homeTeam": "a37f9158-7f82-46bc-908c-c9e2dda7c33b",
-  "outcomes": [],
-  "awayScore": 6,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 7,
-  "statsheet": "a297f78b-b9f5-43c4-b989-f3c9df0f415a",
-  "atBatBalls": 0,
-  "awayBatter": "15ae64cd-f698-4b00-9d61-c9fffd037ae2",
-  "homeBatter": "",
-  "lastUpdate": "Elvis Figueroa gets caught stealing third base.",
-  "awayPitcher": "814bae61-071a-449b-981e-e7afc839d6d6",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "667cb445-c288-4e62-b603-27291c1e475d"
-  ],
-  "homePitcher": "f4a5d734-0ade-4410-abb6-c0cd5a7a1c26",
-  "homeStrikes": 3,
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Philly Pies",
-  "gameComplete": false,
-  "homeTeamName": "Breckenridge Jazz Hands",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#399d8f",
-  "awayTeamEmoji": "0x1F967",
-  "basesOccupied": [
-      0
-  ],
-  "homeTeamColor": "#6388ad",
-  "homeTeamEmoji": "0x1F450",
-  "awayBatterName": "Mickey Woods",
-  "halfInningOuts": 2,
-  "homeBatterName": "",
-  "awayPitcherName": "Ruslan Greatness",
-  "baserunnerCount": 1,
-  "halfInningScore": 3,
-  "homePitcherName": "Agan Harrison",
-  "awayTeamNickname": "Pies",
-  "homeTeamNickname": "Jazz Hands",
-  "awayTeamBatterCount": 32,
-  "homeTeamBatterCount": 26
-};
-
-const sacrificeScore = {
-  _id: 'cbc65532-727b-4ebd-925c-d9f46d4037a0',
-  day: 57,
-  phase: 3,
-  rules: '4ae9d46a-5408-460a-84fb-cbd8d03fff6c',
-  shame: false,
-  inning: 7,
-  season: 2,
-  weather: 11,
-  awayOdds: 0.6395797544362725,
-  awayTeam: 'b72f3061-f573-40d7-832a-5ad475bd7909',
-  homeOdds: 0.3604202455637274,
-  homeTeam: '3f8bbb15-61c0-4e3f-8e4a-907a5fb1565e',
-  outcomes: [],
-  awayScore: 3,
-  finalized: false,
-  gameStart: true,
-  homeScore: 1,
-  statsheet: '16ee8bd7-3bd6-4257-b398-7e8e17a166f3',
-  atBatBalls: 0,
-  awayBatter: '',
-  homeBatter: 'f7715b05-ee69-43e5-a0e5-8e3d34270c82',
-  lastUpdate: 'Helga Burton  scores on the sacrifice.',
-  awayPitcher: '73265ee3-bb35-40d1-b696-1f241a6f5966',
-  awayStrikes: 4,
-  baseRunners: [ '58c9e294-bd49-457c-883f-fb3162fc668e' ],
-  homePitcher: 'dfd5ccbb-90ed-4bfe-83e0-dae9cc763f10',
-  homeStrikes: 3,
-  seriesIndex: 1,
-  terminology: 'b67e9bbb-1495-4e1b-b517-f1444b0a6c8b',
-  topOfInning: true,
-  atBatStrikes: 0,
-  awayTeamName: 'San Francisco Lovers',
-  gameComplete: false,
-  homeTeamName: 'Boston Flowers',
-  isPostseason: false,
-  seriesLength: 3,
-  awayTeamColor: '#780018',
-  awayTeamEmoji: '0x1F48B',
-  basesOccupied: [ 1 ],
-  homeTeamColor: '#f7d1ff',
-  homeTeamEmoji: '0x1F339',
-  awayBatterName: '',
-  halfInningOuts: 2,
-  homeBatterName: 'Caligula Lotus',
-  awayPitcherName: 'Parker Meng',
-  baserunnerCount: 1,
-  halfInningScore: 1,
-  homePitcherName: 'Owen Picklestein',
-  awayTeamNickname: 'Lovers',
-  homeTeamNickname: 'Flowers',
-  awayTeamBatterCount: 28,
-  homeTeamBatterCount: 24
-};
-
-const sacrificeAdvance = {
-  id: '8ac77e5c-fa21-4403-9b67-855d6a469e92',
-  day: 81,
-  phase: 6,
-  rules: '4ae9d46a-5408-460a-84fb-cbd8d03fff6c',
-  shame: false,
-  inning: 7,
-  season: 13,
-  weather: 18,
-  awayOdds: 0.6417733693106469,
-  awayOuts: 3,
-  awayTeam: '747b8e4a-7e50-4638-a973-ea7950a3e739',
-  homeOdds: 0.35822663068935306,
-  homeOuts: 3,
-  homeTeam: '57ec08cc-0411-4643-b304-0e80dbc15ac7',
-  outcomes: [],
-  awayBalls: 4,
-  awayBases: 4,
-  awayScore: 4,
-  finalized: false,
-  gameStart: true,
-  homeBalls: 4,
-  homeBases: 4,
-  homeScore: 2,
-  playCount: 256,
-  stadiumId: '2841dc8c-b5a2-4b3a-a717-9e646e8e29fc',
-  statsheet: '4b567f53-fc67-4f6c-9e11-028357d6d023',
-  atBatBalls: 0,
-  awayBatter: null,
-  homeBatter: null,
-  lastUpdate: 'Paula Mason hit a ground out to Ronan Combs.\n' +
-    'Randy Castillo advances on the sacrifice.',
-  tournament: -1,
-  awayPitcher: '2720559e-9173-4042-aaa0-d3852b72ab2e',
-  awayStrikes: 3,
-  baseRunners: [ 'efafe75e-2f00-4418-914c-9b6675d39264' ],
-  homePitcher: '16aff709-e855-47c8-8818-b9ba66e90fe8',
-  homeStrikes: 3,
-  repeatCount: 0,
-  scoreLedger: '',
-  scoreUpdate: '1 Run scored!',
-  seriesIndex: 1,
-  terminology: 'b67e9bbb-1495-4e1b-b517-f1444b0a6c8b',
-  topOfInning: true,
-  atBatStrikes: 0,
-  awayTeamName: 'Hades Tigers',
-  gameComplete: false,
-  homeTeamName: 'Mexico City Wild Wings',
-  isPostseason: false,
-  isTitleMatch: false,
-  seriesLength: 3,
-  awayBatterMod: '',
-  awayTeamColor: '#5c1c1c',
-  awayTeamEmoji: '0x1F405',
-  basesOccupied: [ 1 ],
-  homeBatterMod: '',
-  homeTeamColor: '#d15700',
-  homeTeamEmoji: '0x1F357',
-  awayBatterName: '',
-  awayPitcherMod: '',
-  baseRunnerMods: [ '' ],
-  gameStartPhase: 18,
-  halfInningOuts: 1,
-  homeBatterName: '',
-  homePitcherMod: '',
-  newInningPhase: -1,
-  topInningScore: 1,
-  awayPitcherName: 'Hiroto Wilcox',
-  baseRunnerNames: [ 'Aldon Cashmoney' ],
-  baserunnerCount: 1,
-  halfInningScore: 1,
-  homePitcherName: 'Mullen Peterson',
-  awayTeamNickname: 'Tigers',
-  homeTeamNickname: 'Wild Wings',
-  secretBaserunner: null,
-  bottomInningScore: 0,
-  awayTeamBatterCount: 33,
-  homeTeamBatterCount: 36,
-  awayTeamSecondaryColor: '#e83622',
-  homeTeamSecondaryColor: '#ee6300'
-};
-
-const endOld = {
-  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
-  "day": 119,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 8,
-  "season": 10,
-  "weather": 14,
-  "awayOdds": 0.38125626427142956,
-  "awayOuts": 3,
-  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
-  "homeOdds": 0.6187437357285706,
-  "homeOuts": 3,
-  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "outcomes": [],
-  "awayBalls": 3,
-  "awayBases": 4,
-  "awayScore": 9,
-  "finalized": true,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 5,
-  "playCount": 306,
-  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Game over.",
-  "tournament": -1,
-  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
-  "awayStrikes": 4,
-  "baseRunners": [],
-  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 5,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hellmouth Sunbeams",
-  "gameComplete": true,
-  "homeTeamName": "Seattle Garages",
-  "isPostseason": true,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#fffbab",
-  "awayTeamEmoji": "0x1F31E",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#2b4075",
-  "homeTeamEmoji": "0x1F3B8",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "awayPitcherName": "Lars Taylor",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Arturo Huerta",
-  "awayTeamNickname": "Sunbeams",
-  "homeTeamNickname": "Garages",
-  "awayTeamBatterCount": 39,
-  "homeTeamBatterCount": 36,
-  "awayTeamSecondaryColor": "#fffbab",
-  "homeTeamSecondaryColor": "#5a83ea"
-};
-
-const endNew = {
-  "id": "a1ad2005-7bb6-4f7f-95bf-8260d43b923e",
-  "day": 115,
-  "phase": 7,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 8,
-  "season": 13,
-  "weather": 18,
-  "awayOdds": 0.3979624149734539,
-  "awayOuts": 3,
-  "awayTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
-  "homeOdds": 0.6020375850265461,
-  "homeOuts": 3,
-  "homeTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
-  "outcomes": [],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 8,
-  "finalized": true,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 5,
-  "playCount": 317,
-  "stadiumId": "36b94380-39a6-4e32-b525-3e888215798a",
-  "statsheet": "76d8535d-e9fc-48ee-9674-6c91dccc0184",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Moist Talkers 8, Tacos 5",
-  "tournament": -1,
-  "awayPitcher": "60026a9d-fc9a-4f5a-94fd-2225398fa3da",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "0bb35615-63f2-4492-80ec-b6b322dc5450",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 2,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Canada Moist Talkers",
-  "gameComplete": true,
-  "homeTeamName": "LA Unlimited Tacos",
-  "isPostseason": true,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#f5feff",
-  "awayTeamEmoji": "0x1F5E3",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#64376e",
-  "homeTeamEmoji": "0x1F32E",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 19,
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Bright Zimmerman",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Sexton Wheerer",
-  "awayTeamNickname": "Moist Talkers",
-  "homeTeamNickname": "Tacos",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 41,
-  "homeTeamBatterCount": 41,
-  "awayTeamSecondaryColor": "#f5feff",
-  "homeTeamSecondaryColor": "#b063c1"
-};
-
-const blooddrain = {
-  "id": "bedb1bdc-2578-4316-bc87-eec44cc11650",
-  "day": 21,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 4,
-  "season": 13,
-  "weather": 9,
-  "awayOdds": 0.5565084195859443,
-  "awayOuts": 3,
-  "awayTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
-  "homeOdds": 0.4434915804140556,
-  "homeOuts": 3,
-  "homeTeam": "979aee4a-6d80-4863-bf1c-ee1a78e06024",
-  "outcomes": [],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 3,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 139,
-  "stadiumId": "e7431679-a45e-408f-92a8-a4db422c4925",
-  "statsheet": "deb7ea31-f38a-492c-8e5a-2a32edc29448",
-  "atBatBalls": 0,
-  "awayBatter": "86d4e22b-f107-4bcf-9625-32d387fcb521",
-  "homeBatter": null,
-  "lastUpdate": "The Blooddrain gurgled!\nSpears Taylor's Siphon activates!\nSpears Taylor siphoned some of York Silk's baserunning ability!\nSpears Taylor adds a Out!",
-  "tournament": -1,
-  "awayPitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "7b55d484-6ea9-4670-8145-986cb9e32412",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Canada Moist Talkers",
-  "gameComplete": false,
-  "homeTeamName": "Hawai'i Fridays",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#f5feff",
-  "awayTeamEmoji": "0x1F5E3",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#3ee652",
-  "homeTeamEmoji": "0x1F3DD",
-  "awayBatterName": "York Silk",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 15,
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Greer Lott",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Stevenson Heat",
-  "awayTeamNickname": "Moist Talkers",
-  "homeTeamNickname": "Fridays",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 19,
-  "homeTeamBatterCount": 14,
-  "awayTeamSecondaryColor": "#f5feff",
-  "homeTeamSecondaryColor": "#3ee652"
-};
-
-const partying = {
-  "id": "3f21fbcf-bc5c-4bb3-9831-82be44f50b23",
-  "day": 97,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 8,
-  "season": 13,
-  "weather": 18,
-  "awayOdds": 0.5473192112165636,
-  "awayOuts": 3,
-  "awayTeam": "b72f3061-f573-40d7-832a-5ad475bd7909",
-  "homeOdds": 0.4526807887834365,
-  "homeOuts": 3,
-  "homeTeam": "46358869-dce9-4a01-bfba-ac24fc56f57e",
-  "outcomes": [
-      "Zoey Kirchner is Partying!",
-      "Bees Taswell is Partying!",
-      "Bottles Suljak is Partying!",
-      "Kelvin Drumsolo is Partying!"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 2,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 4,
-  "playCount": 283,
-  "stadiumId": "031670ae-97a5-4215-b1a1-98e9f1de7c50",
-  "statsheet": "29fe89bb-02e1-444f-8d6f-7619701ac216",
-  "atBatBalls": 0,
-  "awayBatter": "c9e4a49e-e35a-4034-a4c7-293896b40c58",
-  "homeBatter": null,
-  "lastUpdate": "Kelvin Drumsolo is Partying!",
-  "tournament": -1,
-  "awayPitcher": "f2c477fb-28ea-4fcb-943a-9fab22df3da0",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "b629cb7a-414b-4695-a115-7528ab700346",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 2,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "San Francisco Lovers",
-  "gameComplete": false,
-  "homeTeamName": "Core Mechanics",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#780018",
-  "awayTeamEmoji": "0x1F48B",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#d13100",
-  "homeTeamEmoji": "0x1F6E0",
-  "awayBatterName": "Alexander Horne",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 19,
-  "halfInningOuts": 2,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Sandford Garner",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Shirai McElroy",
-  "awayTeamNickname": "Lovers",
-  "homeTeamNickname": "Mechanics",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 32,
-  "homeTeamBatterCount": 31,
-  "awayTeamSecondaryColor": "#da0000",
-  "homeTeamSecondaryColor": "#ff5b29"
-};
-
-const reverb = {
-  "id": "05bc5a39-f72d-44db-9867-b799e01c2f76",
-  "day": 63,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 7,
-  "season": 12,
-  "weather": 13,
-  "awayOdds": 0.43194017300081955,
-  "awayOuts": 3,
-  "awayTeam": "a37f9158-7f82-46bc-908c-c9e2dda7c33b",
-  "homeOdds": 0.5680598269991805,
-  "homeOuts": 3,
-  "homeTeam": "57ec08cc-0411-4643-b304-0e80dbc15ac7",
-  "outcomes": [
-      "The Jazz Hands were shuffled in the Reverb!"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 4.7,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 3,
-  "playCount": 291,
-  "stadiumId": null,
-  "statsheet": "cf8f4de5-4b6d-4790-9ba7-e38f846af795",
-  "atBatBalls": 1,
-  "awayBatter": null,
-  "homeBatter": "2b9f9c25-43ec-4f0b-9937-a5aa23be0d9e",
-  "lastUpdate": "Reverberations are at dangerous levels!\nThe Jazz Hands were shuffled in the Reverb!\nSteph Weeks's Gravity kept them in place!",
-  "tournament": -1,
-  "awayPitcher": "d23a1f7e-0071-444e-8361-6ae01f13036f",
-  "awayStrikes": 4,
-  "baseRunners": [
-      "4542f0b0-3409-4a4a-a9e1-e8e8e5d73fcf"
-  ],
-  "homePitcher": "bd8778e5-02e8-4d1f-9c31-7b63942cc570",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 1,
-  "awayTeamName": "Breckenridge Jazz Hands",
-  "gameComplete": false,
-  "homeTeamName": "Mexico City Wild Wings",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#6388ad",
-  "awayTeamEmoji": "0x1F450",
-  "basesOccupied": [
-      1
-  ],
-  "homeBatterMod": "",
-  "homeTeamColor": "#d15700",
-  "homeTeamEmoji": "0x1F357",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [
-      ""
-  ],
-  "gameStartPhase": 17,
-  "halfInningOuts": 0,
-  "homeBatterName": "Lawrence Horne",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Edric Tosser",
-  "baseRunnerNames": [
-      "Brock Watson"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 0,
-  "homePitcherName": "Cell Barajas",
-  "awayTeamNickname": "Jazz Hands",
-  "homeTeamNickname": "Wild Wings",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 31,
-  "homeTeamBatterCount": 32,
-  "awayTeamSecondaryColor": "#7ba9d7",
-  "homeTeamSecondaryColor": "#ee6300"
-};
-
-const birdsPecked = {
-  "id": "89651db2-c5ba-4746-9568-0b2a40da93d6",
-  "day": 104,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 8,
-  "season": 12,
-  "weather": 11,
-  "awayOdds": 0.4529704358415401,
-  "awayOuts": 3,
-  "awayTeam": "9debc64f-74b7-4ae1-a4d6-fce0144b6ea5",
-  "homeOdds": 0.5470295641584598,
-  "homeOuts": 3,
-  "homeTeam": "ca3f1c8c-c025-4d8e-8eef-5be6accbeb16",
-  "outcomes": [
-      "The Birds pecked Denzel Scott free!"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 4,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 3,
-  "playCount": 288,
-  "stadiumId": "06c991f6-4ef3-4a9a-8254-8a2bc7d3a1e1",
-  "statsheet": "90a4575a-d9a9-410f-a181-8aadca13b76d",
-  "atBatBalls": 1,
-  "awayBatter": "d8ee256f-e3d0-46cb-8c77-b1f88d8c9df9",
-  "homeBatter": null,
-  "lastUpdate": "The Birds circle...\nThe Birds pecked Denzel Scott free!",
-  "tournament": -1,
-  "awayPitcher": "446a3366-3fe3-41bb-bfdd-d8717f2152a9",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "1513aab6-142c-48c6-b43e-fbda65fd64e8",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 1,
-  "awayTeamName": "Houston Spies",
-  "gameComplete": false,
-  "homeTeamName": "Chicago Firefighters",
-  "isPostseason": true,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#67556b",
-  "awayTeamEmoji": "0x1F575",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#8c2a3e",
-  "homeTeamEmoji": "0x1F525",
-  "awayBatterName": "Comfort Septemberish",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 19,
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Marco Escobar",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Caleb Alvarado",
-  "awayTeamNickname": "Spies",
-  "homeTeamNickname": "Firefighters",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 46,
-  "homeTeamBatterCount": 31,
-  "awayTeamSecondaryColor": "#9e82a4",
-  "homeTeamSecondaryColor": "#d13757"
-};
-
-const birdsCircled = {
-  "id": "011a5fe1-85c4-4271-841f-a96ad935e093",
-  "day": 100,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 6,
-  "season": 13,
-  "weather": 11,
-  "awayOdds": 0.5871344288828205,
-  "awayOuts": 3,
-  "awayTeam": "3f8bbb15-61c0-4e3f-8e4a-907a5fb1565e",
-  "homeOdds": 0.41286557111717953,
-  "homeOuts": 3,
-  "homeTeam": "c73b705c-40ad-4633-a6ed-d357ee2e2bcf",
-  "outcomes": [
-      "CONSUMERS ATTACK\nOWEN PICKLESTEIN"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 7,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 2,
-  "playCount": 215,
-  "stadiumId": "7f09624b-b3cf-4493-b768-adf090bebf21",
-  "statsheet": "0eb06702-0aea-4c14-9f00-91db88955fe0",
-  "atBatBalls": 0,
-  "awayBatter": "be18d363-752d-4e4a-b06b-1a7e4641400b",
-  "homeBatter": null,
-  "lastUpdate": "The Birds circle ... but they don't find what they're looking for.",
-  "tournament": -1,
-  "awayPitcher": "c0732e36-3731-4f1a-abdc-daa9563b6506",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "f741dc01-2bae-4459-bfc0-f97536193eea",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 2,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Boston Flowers",
-  "gameComplete": false,
-  "homeTeamName": "Tokyo Lift",
-  "isPostseason": true,
-  "isTitleMatch": false,
-  "seriesLength": 2,
-  "awayBatterMod": "",
-  "awayTeamColor": "#f7d1ff",
-  "awayTeamEmoji": "0x1F339",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#e830ab",
-  "homeTeamEmoji": "\ud83c\udfcb\ufe0f\u200d\u2640\ufe0f",
-  "awayBatterName": "Scores Baserunner",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 17,
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Nagomi Mcdaniel",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Alejandro Leaf",
-  "awayTeamNickname": "Flowers",
-  "homeTeamNickname": "Lift",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 32,
-  "homeTeamBatterCount": 21,
-  "awayTeamSecondaryColor": "#f7d1ff",
-  "homeTeamSecondaryColor": "#e830ab"
-};
-
-const justBirds = {
-  "_id": "902cf51e-22e8-4cfb-b431-35fb8768a55e",
-  "day": 5,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 2,
-  "season": 2,
-  "weather": 11,
-  "awayOdds": 0.7751306680038499,
-  "awayTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
-  "homeOdds": 0.22486933199615022,
-  "homeTeam": "9debc64f-74b7-4ae1-a4d6-fce0144b6ea5",
-  "outcomes": [],
-  "awayScore": 2,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 1,
-  "statsheet": "c378f070-f9f6-4742-9eef-c286d1de7deb",
-  "atBatBalls": 0,
-  "awayBatter": "",
-  "homeBatter": "503a235f-9fa6-41b5-8514-9475c944273f",
-  "lastUpdate": "The birds are mad at you. You specifically. You know who you are.",
-  "awayPitcher": "338694b7-6256-4724-86b6-3884299a5d9e",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "32551e28-3a40-47ae-aed1-ff5bc66be879",
-  "homeStrikes": 3,
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 1,
-  "awayTeamName": "Canada Moist Talkers",
-  "gameComplete": false,
-  "homeTeamName": "Houston Spies",
-  "isPostseason": false,
-  "seriesLength": 3,
-  "awayTeamColor": "#f5feff",
-  "awayTeamEmoji": "0x1F5E3",
-  "basesOccupied": [],
-  "homeTeamColor": "#67556b",
-  "homeTeamEmoji": "0x1F575",
-  "awayBatterName": "",
-  "halfInningOuts": 2,
-  "homeBatterName": "Reese Clark",
-  "awayPitcherName": "PolkaDot Patterson",
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Math Velazquez",
-  "awayTeamNickname": "Moist Talkers",
-  "homeTeamNickname": "Spies",
-  "awayTeamBatterCount": 13,
-  "homeTeamBatterCount": 9
-};
-
-const allergicReaction = {
-  "_id": "327a5b88-e597-4b2a-bdbc-8bf6c96438d8",
-  "day": 19,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 5,
-  "season": 2,
-  "weather": 10,
-  "awayOdds": 0.5370756507699288,
-  "awayTeam": "7966eb04-efcc-499b-8f03-d13916330531",
-  "homeOdds": 0.4629243492300712,
-  "homeTeam": "ca3f1c8c-c025-4d8e-8eef-5be6accbeb16",
-  "outcomes": [
-      "Firefighters hitter Edric Tosser swallowed a stray Peanut and had an allergic reaction!",
-      "Firefighters hitter Joshua Butt swallowed a stray Peanut and had an allergic reaction!"
-  ],
-  "awayScore": 2,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 7,
-  "statsheet": "5146978d-32de-4c0b-9d17-2b1cba339fbe",
-  "atBatBalls": 0,
-  "awayBatter": "4e6ad1a1-7c71-49de-8bd5-c286712faf9e",
-  "homeBatter": "",
-  "lastUpdate": "Firefighters hitter Joshua Butt swallowed a stray Peanut and had an allergic reaction!",
-  "awayPitcher": "2da49de2-34e5-49d0-b752-af2a2ee061be",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "16aff709-e855-47c8-8818-b9ba66e90fe8",
-  "homeStrikes": 3,
-  "seriesIndex": 2,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 1,
-  "awayTeamName": "Yellowstone Magic",
-  "gameComplete": false,
-  "homeTeamName": "Chicago Firefighters",
-  "isPostseason": false,
-  "seriesLength": 3,
-  "awayTeamColor": "#bf0043",
-  "awayTeamEmoji": "0x2728",
-  "basesOccupied": [],
-  "homeTeamColor": "#8c2a3e",
-  "homeTeamEmoji": "0x1F525",
-  "awayBatterName": "Sutton Picklestein",
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "awayPitcherName": "Cory Twelve",
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Mullen Peterson",
-  "awayTeamNickname": "Magic",
-  "homeTeamNickname": "Firefighters",
-  "awayTeamBatterCount": 19,
-  "homeTeamBatterCount": 25
-};
-
-const riv = {
-  "id": "4eee638f-4b07-4471-841d-f50092d25a3c",
-  "day": 86,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 6,
-  "season": 13,
-  "weather": 7,
-  "awayOdds": 0.4274928173200292,
-  "awayOuts": 3,
-  "awayTeam": "7966eb04-efcc-499b-8f03-d13916330531",
-  "homeOdds": 0.5725071826799708,
-  "homeOuts": 3,
-  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
-  "outcomes": [
-      "Rogue Umpire incinerated York Silk!"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 3,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 206,
-  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
-  "statsheet": "c0ce3c76-6a35-4926-a31b-52cf461a6b06",
-  "atBatBalls": 1,
-  "awayBatter": "63512571-2eca-4bc4-8ad9-a5308a22ae22",
-  "homeBatter": null,
-  "lastUpdate": "Rogue Umpire incinerated York Silk!\nThey're replaced by Lucien Patchwork.",
-  "tournament": -1,
-  "awayPitcher": "8adb084b-19fe-4295-bcd2-f92afdb62bd7",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "a3947fbc-50ec-45a4-bca4-49ffebb77dbe",
-      "9a031b9a-16f8-4165-a468-5d0e28a81151"
-  ],
-  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 1,
-  "awayTeamName": "Yellowstone Magic",
-  "gameComplete": false,
-  "homeTeamName": "Canada Moist Talkers",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#bf0043",
-  "awayTeamEmoji": "0x2728",
-  "basesOccupied": [
-      2,
-      1
-  ],
-  "homeBatterMod": "",
-  "homeTeamColor": "#f5feff",
-  "homeTeamEmoji": "0x1F5E3",
-  "awayBatterName": "Oscar Dollie",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [
-      "",
-      ""
-  ],
-  "gameStartPhase": 17,
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Logan Rodriguez",
-  "baseRunnerNames": [
-      "Chorby Short",
-      "Tiana Wheeler"
-  ],
-  "baserunnerCount": 2,
-  "halfInningScore": 0,
-  "homePitcherName": "Greer Lott",
-  "awayTeamNickname": "Magic",
-  "homeTeamNickname": "Moist Talkers",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 26,
-  "homeTeamBatterCount": 22,
-  "awayTeamSecondaryColor": "#f60f63",
-  "homeTeamSecondaryColor": "#f5feff"
-};
-
-const magmatic = {
-  "id": "513eb7eb-1a02-4c11-a376-6c18f508123c",
-  "day": 8,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 4,
-  "season": 13,
-  "weather": 7,
-  "awayOdds": 0.4056945226914672,
-  "awayOuts": 3,
-  "awayTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
-  "homeOdds": 0.5943054773085327,
-  "homeOuts": 3,
-  "homeTeam": "b72f3061-f573-40d7-832a-5ad475bd7909",
-  "outcomes": [],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 3,
-  "playCount": 190,
-  "stadiumId": "1c1614b3-abdb-4008-98eb-12aeb8f7bb54",
-  "statsheet": "eae25e82-4b14-4fe3-8897-1eb8659efa03",
-  "atBatBalls": 2,
-  "awayBatter": null,
-  "homeBatter": "cbd19e6f-3d08-4734-b23f-585330028665",
-  "lastUpdate": "Rogue Umpire tried to incinerate Andrew Solis, but Andrew Solis ate the flame! They became Magmatic!",
-  "tournament": -1,
-  "awayPitcher": "378c07b0-5645-44b5-869f-497d144c7b35",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "c3ae0552-59e8-44bf-ba66-48a96aff35e6",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "New York Millennials",
-  "gameComplete": false,
-  "homeTeamName": "San Francisco Lovers",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#ffd4d8",
-  "awayTeamEmoji": "0x1F4F1",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#780018",
-  "homeTeamEmoji": "0x1F48B",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 15,
-  "halfInningOuts": 2,
-  "homeBatterName": "Knight Triumphant",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Fynn Doyle",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 1,
-  "homePitcherName": "Bontgomery Mullock",
-  "awayTeamNickname": "Millennials",
-  "homeTeamNickname": "Lovers",
-  "secretBaserunner": null,
-  "bottomInningScore": 1,
-  "awayTeamBatterCount": 21,
-  "homeTeamBatterCount": 20,
-  "awayTeamSecondaryColor": "#ffd4d8",
-  "homeTeamSecondaryColor": "#da0000"
-};
-
-const feedback = {
-  "id": "f814e524-a257-4e72-a0c6-f7dae24aa15e",
-  "day": 66,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 4,
-  "season": 13,
-  "weather": 12,
-  "awayOdds": 0.5883245830245686,
-  "awayOuts": 3,
-  "awayTeam": "ca3f1c8c-c025-4d8e-8eef-5be6accbeb16",
-  "homeOdds": 0.4116754169754314,
-  "homeOuts": 3,
-  "homeTeam": "bb4a9de5-c924-4923-a0cb-9d1445f1ee5d",
-  "outcomes": [
-      "NaN and Rivers Rosa switched teams in the feedback!"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 2,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0.7,
-  "playCount": 141,
-  "stadiumId": "8a84154e-80d7-47d5-8f56-295e9a9653d9",
-  "statsheet": "b1bf31cd-96ab-4e94-b858-964b75ba6aa9",
-  "atBatBalls": 0,
-  "awayBatter": "43d5da5f-c6a1-42f1-ab7f-50ea956b6cd5",
-  "homeBatter": null,
-  "lastUpdate": "Reality flickers. Things look different ...\nNaN and Rivers Rosa switch teams in the feedback!\nRivers Rosa is now pitching.",
-  "tournament": -1,
-  "awayPitcher": "e4e4c17d-8128-4704-9e04-f244d4573c4d",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "4bf352d2-6a57-420a-9d45-b23b2b947375",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Chicago Firefighters",
-  "gameComplete": false,
-  "homeTeamName": "Ohio Worms",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#8c2a3e",
-  "awayTeamEmoji": "0x1F525",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#5c4822",
-  "homeTeamEmoji": "0x1F40C",
-  "awayBatterName": "Justice Spoon",
-  "awayPitcherMod": "TRIPLE_THREAT",
-  "baseRunnerMods": [],
-  "gameStartPhase": 15,
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Wesley Poole",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Rivers Rosa",
-  "awayTeamNickname": "Firefighters",
-  "homeTeamNickname": "Worms",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 15,
-  "homeTeamBatterCount": 21,
-  "awayTeamSecondaryColor": "#d13757",
-  "homeTeamSecondaryColor": "#ba9c65"
-};
-
-const zap = {
-  "id": "d3c97762-4ce4-468a-953a-dc77b6ca8500",
-  "day": 18,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 8,
-  "season": 7,
-  "weather": 12,
-  "awayOdds": 0.6210342847132131,
-  "awayTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
-  "homeOdds": 0.37896571528678696,
-  "homeTeam": "b63be8c2-576a-4d6e-8daf-814f8bcea96f",
-  "outcomes": [],
-  "awayBases": 4,
-  "awayScore": 4,
-  "finalized": false,
-  "gameStart": true,
-  "homeBases": 4,
-  "homeScore": 3,
-  "statsheet": "216a63e2-65b4-455b-99f9-528a1c0cd759",
-  "atBatBalls": 2,
-  "awayBatter": null,
-  "homeBatter": "8cf78b49-d0ca-4703-88e8-4bcad26c44b1",
-  "lastUpdate": "The Electricity zaps a strike away!",
-  "awayPitcher": "de21c97e-f575-43b7-8be7-ecc5d8c4eaff",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "9820f2c5-f9da-4a07-b610-c2dd7bee2ef6",
-  "homeStrikes": 3,
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Unlimited Tacos",
-  "gameComplete": false,
-  "homeTeamName": "Miami Dale",
-  "isPostseason": false,
-  "seriesLength": 3,
-  "awayTeamColor": "#64376e",
-  "awayTeamEmoji": "0x1F32E",
-  "basesOccupied": [],
-  "homeTeamColor": "#9141ba",
-  "homeTeamEmoji": "0x1F6A4",
-  "awayBatterName": "",
-  "halfInningOuts": 0,
-  "homeBatterName": "Avila Guzman",
-  "awayPitcherName": "Pitching Machine",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Peanut Bong",
-  "awayTeamNickname": "Tacos",
-  "homeTeamNickname": "Dale",
-  "awayTeamBatterCount": 34,
-  "homeTeamBatterCount": 33
-};
-
-const unstable = {
-  "id": "cd1a7abb-4025-4e27-8e94-382cb87dbe94",
-  "day": 4,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 6,
-  "weather": 7,
-  "awayOdds": 0.40741801592555854,
-  "awayTeam": "adc5b394-8f76-416d-9ce9-813706877b84",
-  "homeOdds": 0.5925819840744415,
-  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "outcomes": [
-      "Jaylen Hotdogfingers hits Dickerson Morse with a pitch! Dickerson Morse is now Unstable!"
-  ],
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 1,
-  "statsheet": "5287b998-deef-4470-9087-a9f56e81e688",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Jaylen Hotdogfingers hits Dickerson Morse with a pitch! Dickerson Morse is now Unstable!",
-  "awayPitcher": "a199a681-decf-4433-b6ab-5454450bbe5e",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "b39b5aae-8571-4c90-887a-6a00f2a2f6fd"
-  ],
-  "homePitcher": "04e14d7b-5021-4250-a3cd-932ba8e0a889",
-  "homeStrikes": 3,
-  "seriesIndex": 2,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Kansas City Breath Mints",
-  "gameComplete": false,
-  "homeTeamName": "Seattle Garages",
-  "isPostseason": false,
-  "seriesLength": 3,
-  "awayTeamColor": "#178f55",
-  "awayTeamEmoji": "0x1F36C",
-  "basesOccupied": [
-      0
-  ],
-  "homeTeamColor": "#2b4075",
-  "homeTeamEmoji": "0x1F3B8",
-  "awayBatterName": "",
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "awayPitcherName": "Leach Ingram",
-  "baseRunnerNames": [
-      "Dickerson Morse"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 0,
-  "homePitcherName": "Jaylen Hotdogfingers",
-  "awayTeamNickname": "Breath Mints",
-  "homeTeamNickname": "Garages",
-  "awayTeamBatterCount": 5,
-  "homeTeamBatterCount": 4
-};
-
-const flickering = {
-  "id": "cd373188-a8b2-48d1-9655-55655df771c6",
-  "day": 99,
-  "phase": 5,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 8,
-  "season": 7,
-  "weather": 12,
-  "awayOdds": 0.40508966169029764,
-  "awayTeam": "23e4cbc1-e9cd-47fa-a35b-bfa06f726cb7",
-  "homeOdds": 0.5949103383097023,
-  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "outcomes": [
-      "Jaylen Hotdogfingers hits Jaxon Buckley with a pitch! Jaxon Buckley is now Flickering!",
-      "Jaylen Hotdogfingers hits Summers Pony with a pitch! Summers Pony is now Flickering!"
-  ],
-  "awayBases": 4,
-  "awayScore": 7,
-  "finalized": false,
-  "gameStart": true,
-  "homeBases": 4,
-  "homeScore": 3,
-  "statsheet": "962cca60-0967-4545-a2e6-e512a3e9e2e2",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Jaylen Hotdogfingers hits Summers Pony with a pitch! Summers Pony is now Flickering!",
-  "awayPitcher": "1732e623-ffc2-40f0-87ba-fdcf97131f1f",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "81a0889a-4606-4f49-b419-866b57331383"
-  ],
-  "homePitcher": "04e14d7b-5021-4250-a3cd-932ba8e0a889",
-  "homeStrikes": 3,
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Philly Pies",
-  "gameComplete": false,
-  "homeTeamName": "Seattle Garages",
-  "isPostseason": true,
-  "seriesLength": 5,
-  "awayTeamColor": "#399d8f",
-  "awayTeamEmoji": "0x1F967",
-  "basesOccupied": [
-      0
-  ],
-  "homeTeamColor": "#2b4075",
-  "homeTeamEmoji": "0x1F3B8",
-  "awayBatterName": "",
-  "halfInningOuts": 2,
-  "homeBatterName": "",
-  "awayPitcherName": "Betsy Trombone",
-  "baseRunnerNames": [
-      "Summers Pony"
-  ],
-  "baserunnerCount": 1,
-  "halfInningScore": 2,
-  "homePitcherName": "Jaylen Hotdogfingers",
-  "awayTeamNickname": "Pies",
-  "homeTeamNickname": "Garages",
-  "awayTeamBatterCount": 39,
-  "homeTeamBatterCount": 30
-};
-
-const consumers = {
-  "id": "5d02a9f5-6b0f-464b-8dfe-71361f7af95e",
-  "day": 111,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 13,
-  "weather": 10,
-  "awayOdds": 0.4363791114405691,
-  "awayOuts": 3,
-  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
-  "homeOdds": 0.5636208885594308,
-  "homeOuts": 3,
-  "homeTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
-  "outcomes": [
-      "CONSUMERS ATTACK\nHIROTO WILCOX"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 14,
-  "stadiumId": "36b94380-39a6-4e32-b525-3e888215798a",
-  "statsheet": "e6227e9d-7447-478e-b134-78fe49280e63",
-  "atBatBalls": 1,
-  "awayBatter": null,
-  "homeBatter": "03b80a57-77ea-4913-9be4-7a85c3594745",
-  "lastUpdate": "CONSUMERS ATTACK\nHIROTO WILCOX",
-  "tournament": -1,
-  "awayPitcher": "1db2f602-64b1-4a5c-8697-1932cc2c6df1",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "d81ce662-07b6-4a73-baa4-acbbb41f9dc5",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 4,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Hades Tigers",
-  "gameComplete": false,
-  "homeTeamName": "LA Unlimited Tacos",
-  "isPostseason": true,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#5c1c1c",
-  "awayTeamEmoji": "0x1F405",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#64376e",
-  "homeTeamEmoji": "0x1F32E",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 11,
-  "halfInningOuts": 0,
-  "homeBatterName": "Halexandrey Walton",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Mummy Melcon",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Yummy Elliott",
-  "awayTeamNickname": "Tigers",
-  "homeTeamNickname": "Tacos",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 2,
-  "homeTeamBatterCount": 0,
-  "awayTeamSecondaryColor": "#e83622",
-  "homeTeamSecondaryColor": "#b063c1"
-};
-
-const grandSlam = {
-  "id": "a0d0fb26-3aa2-4af2-99fe-08b4a34f439c",
-  "day": 87,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 13,
-  "weather": 18,
-  "awayOdds": 0.4721566362068745,
-  "awayOuts": 3,
-  "awayTeam": "d9f89a8a-c563-493e-9d64-78e4f9a55d4a",
-  "homeOdds": 0.5278433637931255,
-  "homeOuts": 3,
-  "homeTeam": "a37f9158-7f82-46bc-908c-c9e2dda7c33b",
-  "outcomes": [],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 4,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 27,
-  "stadiumId": "60e78a6c-fe49-4283-8293-735e9fa7e3e9",
-  "statsheet": "55f95f79-d215-4e69-9bb4-267240572621",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "Yurts Buttercup hits a grand slam!",
-  "tournament": -1,
-  "awayPitcher": "0ef1bc34-64ee-4c7b-8be2-88322b407a58",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "d23a1f7e-0071-444e-8361-6ae01f13036f",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "4 Runs scored!",
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "Atlantis Georgias",
-  "gameComplete": false,
-  "homeTeamName": "Breckenridge Jazz Hands",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#00dbc2",
-  "awayTeamEmoji": "0x1F531",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#6388ad",
-  "homeTeamEmoji": "0x1F450",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 11,
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 4,
-  "awayPitcherName": "Geraldine Frost",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 4,
-  "homePitcherName": "Edric Tosser",
-  "awayTeamNickname": "Georgias",
-  "homeTeamNickname": "Jazz Hands",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 4,
-  "homeTeamBatterCount": -1,
-  "awayTeamSecondaryColor": "#5cffec",
-  "homeTeamSecondaryColor": "#7ba9d7"
-};
-
-const grandSlamPre12 = {
-  "_id": "0bdca458-2796-4011-949d-c0e6faa03def",
-  "day": 24,
-  "phase": 3,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 2,
-  "weather": 11,
-  "awayOdds": 0.3790636711401531,
-  "awayTeam": "23e4cbc1-e9cd-47fa-a35b-bfa06f726cb7",
-  "homeOdds": 0.620936328859847,
-  "homeTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
-  "outcomes": [],
-  "awayScore": 2,
-  "finalized": false,
-  "gameStart": true,
-  "homeScore": 4,
-  "statsheet": "ef1c2c65-a398-4d68-87d6-1522918a1027",
-  "atBatBalls": 0,
-  "awayBatter": "",
-  "homeBatter": "",
-  "lastUpdate": "Alexander Horne hits a grand slam!",
-  "awayPitcher": "814bae61-071a-449b-981e-e7afc839d6d6",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "333067fd-c2b4-4045-a9a4-e87a8d0332d0",
-  "homeStrikes": 4,
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Philly Pies",
-  "gameComplete": false,
-  "homeTeamName": "Hellmouth Sunbeams",
-  "isPostseason": false,
-  "seriesLength": 3,
-  "awayTeamColor": "#399d8f",
-  "awayTeamEmoji": "0x1F967",
-  "basesOccupied": [],
-  "homeTeamColor": "#fffbab",
-  "homeTeamEmoji": "0x1F31E",
-  "awayBatterName": "",
-  "halfInningOuts": 2,
-  "homeBatterName": "",
-  "awayPitcherName": "Ruslan Greatness",
-  "baserunnerCount": 0,
-  "halfInningScore": 4,
-  "homePitcherName": "Miguel James",
-  "awayTeamNickname": "Pies",
-  "homeTeamNickname": "Sunbeams",
-  "awayTeamBatterCount": 5,
-  "homeTeamBatterCount": 5
-};
-
-const maximumBlaseball = {
-  "id": "0b4a26a8-46d3-4d2c-a6eb-5e1d3682f1df",
-  "day": 96,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 13,
-  "weather": 1,
-  "awayOdds": 0.29425858188999554,
-  "awayOuts": 3,
-  "awayTeam": "bfd38797-8404-4b38-8b82-341da28b1f83",
-  "homeOdds": 0.7057414181100045,
-  "homeOuts": 3,
-  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
-  "outcomes": [
-      "Esme Ramsey is Partying!"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 6,
-  "playCount": 115,
-  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
-  "statsheet": "8d903046-8831-4b4e-9d98-49b590e40591",
-  "atBatBalls": 3,
-  "awayBatter": null,
-  "homeBatter": "9f85676a-7411-444a-8ae2-c7f8f73c285c",
-  "lastUpdate": "Esme Ramsey is Partying!",
-  "tournament": -1,
-  "awayPitcher": "5b5bcc6c-d011-490f-b084-6fdc2c52f958",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "c31d874c-1b4d-40f2-a1b3-42542e934047",
-      "a7d8196a-ca6b-4dab-a9d7-c27f3e86cc21",
-      "611d18e0-b972-4cdd-afc2-793c56bfe5a9"
-  ],
-  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 2,
-  "awayTeamName": "Charleston Shoe Thieves",
-  "gameComplete": false,
-  "homeTeamName": "Canada Moist Talkers",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#ffce0a",
-  "awayTeamEmoji": "0x1F45F",
-  "basesOccupied": [
-      2,
-      1,
-      0
-  ],
-  "homeBatterMod": "",
-  "homeTeamColor": "#f5feff",
-  "homeTeamEmoji": "0x1F5E3",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [
-      "",
-      "",
-      ""
-  ],
-  "gameStartPhase": 12,
-  "halfInningOuts": 2,
-  "homeBatterName": "Lachlan Shelton",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Simba Davis",
-  "baseRunnerNames": [
-      "Cedric Spliff",
-      "Commissioner Vapor",
-      "Alston Cerveza"
-  ],
-  "baserunnerCount": 3,
-  "halfInningScore": 0,
-  "homePitcherName": "Greer Lott",
-  "awayTeamNickname": "Shoe Thieves",
-  "homeTeamNickname": "Moist Talkers",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 7,
-  "homeTeamBatterCount": 17,
-  "awayTeamSecondaryColor": "#ffce0a",
-  "homeTeamSecondaryColor": "#f5feff"
-};
-
-const salmon = {
-  "id": "8556e3be-b432-4877-a286-499ed9a27bcd",
-  "day": 0,
-  "phase": 2,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 3,
-  "season": 14,
-  "weather": 19,
-  "awayOdds": 0.4164700715377954,
-  "awayOuts": 3,
-  "awayTeam": "b63be8c2-576a-4d6e-8daf-814f8bcea96f",
-  "homeOdds": 0.5835299284622046,
-  "homeOuts": 3,
-  "homeTeam": "d9f89a8a-c563-493e-9d64-78e4f9a55d4a",
-  "outcomes": [],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 1,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 3,
-  "playCount": 166,
-  "stadiumId": "37c2743f-1a15-4e3f-bdb1-a6461df93434",
-  "statsheet": "2c3ba5aa-370f-423b-b19a-ae6afc70c493",
-  "atBatBalls": 0,
-  "awayBatter": null,
-  "homeBatter": null,
-  "lastUpdate": "The Salmon swim upstream!\nInning 5 begins again.\nNo Runs are lost.",
-  "tournament": -1,
-  "awayPitcher": "a38ada0a-aeac-4a3d-b9a5-968687ccd2f9",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "bc29afc1-c954-4def-978b-a59ae5def3c3",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": false,
-  "atBatStrikes": 0,
-  "awayTeamName": "Miami Dale",
-  "gameComplete": false,
-  "homeTeamName": "Atlantis Georgias",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#9141ba",
-  "awayTeamEmoji": "0x1F6A4",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#00dbc2",
-  "homeTeamEmoji": "0x1F531",
-  "awayBatterName": "",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 18,
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": 0,
-  "topInningScore": 0,
-  "awayPitcherName": "Sixpack Santiago",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Rigby Friedrich",
-  "awayTeamNickname": "Dale",
-  "homeTeamNickname": "Georgias",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 18,
-  "homeTeamBatterCount": 19,
-  "awayTeamSecondaryColor": "#cd76fa",
-  "homeTeamSecondaryColor": "#5cffec"
-};
-
-const consumersDefend = {
-  "id": "0e07ec78-9ee2-4fd1-842c-d0b9a8cd948f",
-  "day": 0,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 1,
-  "season": 15,
-  "weather": 11,
-  "awayOdds": 0.4855122940914459,
-  "awayOuts": 3,
-  "awayTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
-  "homeOdds": 0.514487705908554,
-  "homeOuts": 3,
-  "homeTeam": "8d87c468-699a-47a8-b40d-cfb73a5660ad",
-  "outcomes": [
-      "CONSUMERS ATTACK\nTOT CLARK DEFENDS\n\nBAT BREAKS",
-      "CONSUMERS ATTACK\nPENELOPE MATHEWS DEFENDS\n\nBAT BREAKS"
-  ],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 1,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 1,
-  "playCount": 63,
-  "stadiumId": "cfb57d7c-4118-4b0a-85cc-4e3a51a66cb6",
-  "statsheet": "4de2bfc7-5b97-407b-9ec5-9e9c884dda4a",
-  "atBatBalls": 0,
-  "awayBatter": "1c73f91e-0562-480d-9543-2aab1d5e5acd",
-  "homeBatter": null,
-  "lastUpdate": "CONSUMERS ATTACK\nPENELOPE MATHEWS DEFENDS\n\nBAT BREAKS",
-  "tournament": -1,
-  "awayPitcher": "1732e623-ffc2-40f0-87ba-fdcf97131f1f",
-  "awayStrikes": 3,
-  "baseRunners": [
-      "8b53ce82-4b1a-48f0-999d-1774b3719202",
-      "f3ddfd87-73a2-4681-96fe-829476c97886",
-      "c18961e9-ef3f-4954-bd6b-9fe01c615186"
-  ],
-  "homePitcher": "7dcf6902-632f-48c5-936a-7cf88802b93a",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 1,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 2,
-  "awayTeamName": "Seattle Garages",
-  "gameComplete": false,
-  "homeTeamName": "Baltimore Crabs",
-  "isPostseason": false,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#2b4075",
-  "awayTeamEmoji": "0x1F3B8",
-  "basesOccupied": [
-      2,
-      1,
-      0
-  ],
-  "homeBatterMod": "",
-  "homeTeamColor": "#593037",
-  "homeTeamEmoji": "0x1F980",
-  "awayBatterName": "Sparks Beans",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [
-      "",
-      "",
-      ""
-  ],
-  "gameStartPhase": 17,
-  "halfInningOuts": 0,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 1,
-  "awayPitcherName": "Betsy Trombone",
-  "baseRunnerNames": [
-      "Oliver Mueller",
-      "Theodore Duende",
-      "Carmelo Plums"
-  ],
-  "baserunnerCount": 3,
-  "halfInningScore": 1,
-  "homePitcherName": "Parker Parra",
-  "awayTeamNickname": "Garages",
-  "homeTeamNickname": "Crabs",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 8,
-  "homeTeamBatterCount": 3,
-  "awayTeamSecondaryColor": "#5a83ea",
-  "homeTeamSecondaryColor": "#b05c6b"
-};
-
-const batter = {
-  "id": "b2cafd20-a799-48f6-abd7-c99bd79a6bd1",
-  "day": 116,
-  "phase": 6,
-  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
-  "shame": false,
-  "inning": 0,
-  "season": 13,
-  "weather": 18,
-  "awayOdds": 0.44891216386293564,
-  "awayOuts": 3,
-  "awayTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
-  "homeOdds": 0.5510878361370642,
-  "homeOuts": 3,
-  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
-  "outcomes": [],
-  "awayBalls": 4,
-  "awayBases": 4,
-  "awayScore": 0,
-  "finalized": false,
-  "gameStart": true,
-  "homeBalls": 4,
-  "homeBases": 4,
-  "homeScore": 0,
-  "playCount": 11,
-  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
-  "statsheet": "a1d470e8-8f19-4186-a64b-73f0195a896f",
-  "atBatBalls": 0,
-  "awayBatter": "63df8701-1871-4987-87d7-b55d4f1df2e9",
-  "homeBatter": null,
-  "lastUpdate": "M---w-ll Mason batting for the Tacos.",
-  "tournament": -1,
-  "awayPitcher": "d81ce662-07b6-4a73-baa4-acbbb41f9dc5",
-  "awayStrikes": 3,
-  "baseRunners": [],
-  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
-  "homeStrikes": 3,
-  "repeatCount": 0,
-  "scoreLedger": "",
-  "scoreUpdate": "",
-  "seriesIndex": 3,
-  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
-  "topOfInning": true,
-  "atBatStrikes": 0,
-  "awayTeamName": "LA Unlimited Tacos",
-  "gameComplete": false,
-  "homeTeamName": "Canada Moist Talkers",
-  "isPostseason": true,
-  "isTitleMatch": false,
-  "seriesLength": 3,
-  "awayBatterMod": "",
-  "awayTeamColor": "#64376e",
-  "awayTeamEmoji": "0x1F32E",
-  "basesOccupied": [],
-  "homeBatterMod": "",
-  "homeTeamColor": "#f5feff",
-  "homeTeamEmoji": "0x1F5E3",
-  "awayBatterName": "M---w-ll Mason",
-  "awayPitcherMod": "",
-  "baseRunnerMods": [],
-  "gameStartPhase": 11,
-  "halfInningOuts": 1,
-  "homeBatterName": "",
-  "homePitcherMod": "",
-  "newInningPhase": -1,
-  "topInningScore": 0,
-  "awayPitcherName": "Yummy Elliott",
-  "baseRunnerNames": [],
-  "baserunnerCount": 0,
-  "halfInningScore": 0,
-  "homePitcherName": "Greer Lott",
-  "awayTeamNickname": "Tacos",
-  "homeTeamNickname": "Moist Talkers",
-  "secretBaserunner": null,
-  "bottomInningScore": 0,
-  "awayTeamBatterCount": 1,
-  "homeTeamBatterCount": -1,
-  "awayTeamSecondaryColor": "#b063c1",
-  "homeTeamSecondaryColor": "#f5feff"
 };
 
 module.exports = {
   noData,
-  beforeFirstPitch,
-  flyout,
-  groundOut,
-  strikeOut,
-  lastOut: groundOut,
-  secondHalf,
-  firstHalf,
-  hit,
-  single: hit,
-  dbl,
-  homeRun,
-  bigBucket: homeRun,
-  steal,
-  stealScore: steal,
-  soloHR,
-  twoRunHR,
-  triple,
-  hitButNoScore,
-  stealScore,
-  stealSecond,
-  walk,
-  walkScores,
-  stealHome,
-  caughtStealing,
-  sacrificeScore,
-  sacrificeAdvance,
-  endOld,
-  endNew,
-  blooddrain,
-  partying,
-  reverb,
-  birdsPecked,
-  birdsCircled,
-  justBirds,
-  allergicReaction,
-  riv,
-  magmatic,
-  feedback,
-  zap,
-  unstable,
-  flickering,
-  consumers,
-  grandSlam,
-  grandSlamPre12,
-  maximumBlaseball,
-  salmon,
-  consumersDefend,
-  batter,
-  batterAlready: grandSlam,
-  runnerFirst: hit,
-  runners1and2: stealHome,
+  ...GAME_DATA,
 };
