+++ test/data/hits.js
@@ -1,0 +1,649 @@
+const hit = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 0,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Conrad Vaughan hits a Single!",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [
+      "5dbf11c0-994a-4482-bd1e-99379148ee45"
+  ],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [
+      0
+  ],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "",
+  "halfInningOuts": 2,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [
+      "Conrad Vaughan"
+  ],
+  "baserunnerCount": 1,
+  "halfInningScore": 0,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": 2,
+  "homeTeamBatterCount": 2
+};
+
+const dbl = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 1,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Ren Morin hits a Double! 1 scores.",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [
+      "32c9bce6-6e52-40fa-9f64-3629b3d026a8"
+  ],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [
+      1
+  ],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "",
+  "halfInningOuts": 2,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [
+      "Ren Morin"
+  ],
+  "baserunnerCount": 1,
+  "halfInningScore": 1,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": 6,
+  "homeTeamBatterCount": 3
+};
+
+const homeRun = {
+  "id": "b2cafd20-a799-48f6-abd7-c99bd79a6bd1",
+  "day": 116,
+  "phase": 6,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 13,
+  "weather": 18,
+  "awayOdds": 0.44891216386293564,
+  "awayOuts": 3,
+  "awayTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
+  "homeOdds": 0.5510878361370642,
+  "homeOuts": 3,
+  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
+  "outcomes": [],
+  "awayBalls": 4,
+  "awayBases": 4,
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 3,
+  "playCount": 65,
+  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
+  "statsheet": "a1d470e8-8f19-4186-a64b-73f0195a896f",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Alston Cerveza hits a 2-run home run!\nThe ball lands in a Big Bucket. An extra Run scores!",
+  "tournament": -1,
+  "awayPitcher": "d81ce662-07b6-4a73-baa4-acbbb41f9dc5",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "scoreLedger": "",
+  "scoreUpdate": "3 Runs scored!",
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "LA Unlimited Tacos",
+  "gameComplete": false,
+  "homeTeamName": "Canada Moist Talkers",
+  "isPostseason": true,
+  "isTitleMatch": false,
+  "seriesLength": 3,
+  "awayBatterMod": "",
+  "awayTeamColor": "#64376e",
+  "awayTeamEmoji": "0x1F32E",
+  "basesOccupied": [],
+  "homeBatterMod": "",
+  "homeTeamColor": "#f5feff",
+  "homeTeamEmoji": "0x1F5E3",
+  "awayBatterName": "",
+  "awayPitcherMod": "",
+  "baseRunnerMods": [],
+  "gameStartPhase": 12,
+  "halfInningOuts": 1,
+  "homeBatterName": "",
+  "homePitcherMod": "",
+  "newInningPhase": -1,
+  "topInningScore": 0,
+  "awayPitcherName": "Yummy Elliott",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 3,
+  "homePitcherName": "Greer Lott",
+  "awayTeamNickname": "Tacos",
+  "homeTeamNickname": "Moist Talkers",
+  "secretBaserunner": null,
+  "bottomInningScore": 3,
+  "awayTeamBatterCount": 8,
+  "homeTeamBatterCount": 6,
+  "awayTeamSecondaryColor": "#b063c1",
+  "homeTeamSecondaryColor": "#f5feff"
+};
+
+const soloHR = {
+  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
+  "day": 119,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 10,
+  "weather": 14,
+  "awayOdds": 0.38125626427142956,
+  "awayOuts": 3,
+  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
+  "homeOdds": 0.6187437357285706,
+  "homeOuts": 3,
+  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
+  "outcomes": [],
+  "awayBalls": 3,
+  "awayBases": 4,
+  "awayScore": 4,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 0,
+  "playCount": 48,
+  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Hahn Fox hits a solo home run!",
+  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
+  "awayStrikes": 4,
+  "baseRunners": [],
+  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "seriesIndex": 5,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hellmouth Sunbeams",
+  "gameComplete": false,
+  "homeTeamName": "Seattle Garages",
+  "isPostseason": true,
+  "seriesLength": 3,
+  "awayTeamColor": "#fffbab",
+  "awayTeamEmoji": "0x1F31E",
+  "basesOccupied": [],
+  "homeTeamColor": "#2b4075",
+  "homeTeamEmoji": "0x1F3B8",
+  "awayBatterName": "",
+  "halfInningOuts": 2,
+  "homeBatterName": "",
+  "awayPitcherName": "Lars Taylor",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 4,
+  "homePitcherName": "Arturo Huerta",
+  "awayTeamNickname": "Sunbeams",
+  "homeTeamNickname": "Garages",
+  "awayTeamBatterCount": 8,
+  "homeTeamBatterCount": 3,
+  "awayTeamSecondaryColor": "#fffbab",
+  "homeTeamSecondaryColor": "#5a83ea"
+};
+
+const twoRunHR = {
+  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
+  "day": 119,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 10,
+  "weather": 14,
+  "awayOdds": 0.38125626427142956,
+  "awayOuts": 3,
+  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
+  "homeOdds": 0.6187437357285706,
+  "homeOuts": 3,
+  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
+  "outcomes": [],
+  "awayBalls": 3,
+  "awayBases": 4,
+  "awayScore": 4,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 2,
+  "playCount": 64,
+  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Oliver Mueller hits a 2-run home run!",
+  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
+  "awayStrikes": 4,
+  "baseRunners": [],
+  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "seriesIndex": 5,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hellmouth Sunbeams",
+  "gameComplete": false,
+  "homeTeamName": "Seattle Garages",
+  "isPostseason": true,
+  "seriesLength": 3,
+  "awayTeamColor": "#fffbab",
+  "awayTeamEmoji": "0x1F31E",
+  "basesOccupied": [],
+  "homeTeamColor": "#2b4075",
+  "homeTeamEmoji": "0x1F3B8",
+  "awayBatterName": "",
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "awayPitcherName": "Lars Taylor",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 2,
+  "homePitcherName": "Arturo Huerta",
+  "awayTeamNickname": "Sunbeams",
+  "homeTeamNickname": "Garages",
+  "awayTeamBatterCount": 9,
+  "homeTeamBatterCount": 5,
+  "awayTeamSecondaryColor": "#fffbab",
+  "homeTeamSecondaryColor": "#5a83ea"
+};
+
+const triple = {
+  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
+  "day": 119,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 10,
+  "weather": 14,
+  "awayOdds": 0.38125626427142956,
+  "awayOuts": 3,
+  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
+  "homeOdds": 0.6187437357285706,
+  "homeOuts": 3,
+  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
+  "outcomes": [],
+  "awayBalls": 3,
+  "awayBases": 4,
+  "awayScore": 4,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 0,
+  "playCount": 60,
+  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Summers Pony hits a Triple!",
+  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
+  "awayStrikes": 4,
+  "baseRunners": [
+      "81a0889a-4606-4f49-b419-866b57331383"
+  ],
+  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "seriesIndex": 5,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hellmouth Sunbeams",
+  "gameComplete": false,
+  "homeTeamName": "Seattle Garages",
+  "isPostseason": true,
+  "seriesLength": 3,
+  "awayTeamColor": "#fffbab",
+  "awayTeamEmoji": "0x1F31E",
+  "basesOccupied": [
+      2
+  ],
+  "homeTeamColor": "#2b4075",
+  "homeTeamEmoji": "0x1F3B8",
+  "awayBatterName": "",
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "awayPitcherName": "Lars Taylor",
+  "baseRunnerNames": [
+      "Summers Pony"
+  ],
+  "baserunnerCount": 1,
+  "halfInningScore": 0,
+  "homePitcherName": "Arturo Huerta",
+  "awayTeamNickname": "Sunbeams",
+  "homeTeamNickname": "Garages",
+  "awayTeamBatterCount": 9,
+  "homeTeamBatterCount": 4,
+  "awayTeamSecondaryColor": "#fffbab",
+  "homeTeamSecondaryColor": "#5a83ea"
+};
+
+const hitButNoScore = {
+  "id": "b2cafd20-a799-48f6-abd7-c99bd79a6bd1",
+  "day": 116,
+  "phase": 6,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 13,
+  "weather": 18,
+  "awayOdds": 0.44891216386293564,
+  "awayOuts": 3,
+  "awayTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
+  "homeOdds": 0.5510878361370642,
+  "homeOuts": 3,
+  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
+  "outcomes": [],
+  "awayBalls": 4,
+  "awayBases": 4,
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 3,
+  "playCount": 73,
+  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
+  "statsheet": "a1d470e8-8f19-4186-a64b-73f0195a896f",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Lachlan Shelton hits a Single!",
+  "tournament": -1,
+  "awayPitcher": "d81ce662-07b6-4a73-baa4-acbbb41f9dc5",
+  "awayStrikes": 3,
+  "baseRunners": [
+      "9f85676a-7411-444a-8ae2-c7f8f73c285c"
+  ],
+  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "scoreLedger": "",
+  "scoreUpdate": "",
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "LA Unlimited Tacos",
+  "gameComplete": false,
+  "homeTeamName": "Canada Moist Talkers",
+  "isPostseason": true,
+  "isTitleMatch": false,
+  "seriesLength": 3,
+  "awayBatterMod": "",
+  "awayTeamColor": "#64376e",
+  "awayTeamEmoji": "0x1F32E",
+  "basesOccupied": [
+      0
+  ],
+  "homeBatterMod": "",
+  "homeTeamColor": "#f5feff",
+  "homeTeamEmoji": "0x1F5E3",
+  "awayBatterName": "",
+  "awayPitcherMod": "",
+  "baseRunnerMods": [
+      ""
+  ],
+  "gameStartPhase": 12,
+  "halfInningOuts": 1,
+  "homeBatterName": "",
+  "homePitcherMod": "",
+  "newInningPhase": -1,
+  "topInningScore": 0,
+  "awayPitcherName": "Yummy Elliott",
+  "baseRunnerNames": [
+      "Lachlan Shelton"
+  ],
+  "baserunnerCount": 1,
+  "halfInningScore": 3,
+  "homePitcherName": "Greer Lott",
+  "awayTeamNickname": "Tacos",
+  "homeTeamNickname": "Moist Talkers",
+  "secretBaserunner": null,
+  "bottomInningScore": 3,
+  "awayTeamBatterCount": 8,
+  "homeTeamBatterCount": 7,
+  "awayTeamSecondaryColor": "#b063c1",
+  "homeTeamSecondaryColor": "#f5feff"
+};
+
+const grandSlam = {
+  "id": "a0d0fb26-3aa2-4af2-99fe-08b4a34f439c",
+  "day": 87,
+  "phase": 6,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 0,
+  "season": 13,
+  "weather": 18,
+  "awayOdds": 0.4721566362068745,
+  "awayOuts": 3,
+  "awayTeam": "d9f89a8a-c563-493e-9d64-78e4f9a55d4a",
+  "homeOdds": 0.5278433637931255,
+  "homeOuts": 3,
+  "homeTeam": "a37f9158-7f82-46bc-908c-c9e2dda7c33b",
+  "outcomes": [],
+  "awayBalls": 4,
+  "awayBases": 4,
+  "awayScore": 4,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 0,
+  "playCount": 27,
+  "stadiumId": "60e78a6c-fe49-4283-8293-735e9fa7e3e9",
+  "statsheet": "55f95f79-d215-4e69-9bb4-267240572621",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Yurts Buttercup hits a grand slam!",
+  "tournament": -1,
+  "awayPitcher": "0ef1bc34-64ee-4c7b-8be2-88322b407a58",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "d23a1f7e-0071-444e-8361-6ae01f13036f",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "scoreLedger": "",
+  "scoreUpdate": "4 Runs scored!",
+  "seriesIndex": 1,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Atlantis Georgias",
+  "gameComplete": false,
+  "homeTeamName": "Breckenridge Jazz Hands",
+  "isPostseason": false,
+  "isTitleMatch": false,
+  "seriesLength": 3,
+  "awayBatterMod": "",
+  "awayTeamColor": "#00dbc2",
+  "awayTeamEmoji": "0x1F531",
+  "basesOccupied": [],
+  "homeBatterMod": "",
+  "homeTeamColor": "#6388ad",
+  "homeTeamEmoji": "0x1F450",
+  "awayBatterName": "",
+  "awayPitcherMod": "",
+  "baseRunnerMods": [],
+  "gameStartPhase": 11,
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "homePitcherMod": "",
+  "newInningPhase": -1,
+  "topInningScore": 4,
+  "awayPitcherName": "Geraldine Frost",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 4,
+  "homePitcherName": "Edric Tosser",
+  "awayTeamNickname": "Georgias",
+  "homeTeamNickname": "Jazz Hands",
+  "secretBaserunner": null,
+  "bottomInningScore": 0,
+  "awayTeamBatterCount": 4,
+  "homeTeamBatterCount": -1,
+  "awayTeamSecondaryColor": "#5cffec",
+  "homeTeamSecondaryColor": "#7ba9d7"
+};
+
+const grandSlamPre12 = {
+  "_id": "0bdca458-2796-4011-949d-c0e6faa03def",
+  "day": 24,
+  "phase": 3,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 0,
+  "season": 2,
+  "weather": 11,
+  "awayOdds": 0.3790636711401531,
+  "awayTeam": "23e4cbc1-e9cd-47fa-a35b-bfa06f726cb7",
+  "homeOdds": 0.620936328859847,
+  "homeTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
+  "outcomes": [],
+  "awayScore": 2,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 4,
+  "statsheet": "ef1c2c65-a398-4d68-87d6-1522918a1027",
+  "atBatBalls": 0,
+  "awayBatter": "",
+  "homeBatter": "",
+  "lastUpdate": "Alexander Horne hits a grand slam!",
+  "awayPitcher": "814bae61-071a-449b-981e-e7afc839d6d6",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "333067fd-c2b4-4045-a9a4-e87a8d0332d0",
+  "homeStrikes": 4,
+  "seriesIndex": 1,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "Philly Pies",
+  "gameComplete": false,
+  "homeTeamName": "Hellmouth Sunbeams",
+  "isPostseason": false,
+  "seriesLength": 3,
+  "awayTeamColor": "#399d8f",
+  "awayTeamEmoji": "0x1F967",
+  "basesOccupied": [],
+  "homeTeamColor": "#fffbab",
+  "homeTeamEmoji": "0x1F31E",
+  "awayBatterName": "",
+  "halfInningOuts": 2,
+  "homeBatterName": "",
+  "awayPitcherName": "Ruslan Greatness",
+  "baserunnerCount": 0,
+  "halfInningScore": 4,
+  "homePitcherName": "Miguel James",
+  "awayTeamNickname": "Pies",
+  "homeTeamNickname": "Sunbeams",
+  "awayTeamBatterCount": 5,
+  "homeTeamBatterCount": 5
+};
+
+
+module.exports = {
+  hit,
+  single: hit,
+  dbl,
+  homeRun,
+  bigBucket: homeRun,
+  soloHR,
+  twoRunHR,
+  triple,
+  hitButNoScore,
+  grandSlam,
+  grandSlamPre12,
+};
+
