+++ test/data/misc.js
@@ -1,0 +1,187 @@
+const maximumBlaseball = {
+  "id": "0b4a26a8-46d3-4d2c-a6eb-5e1d3682f1df",
+  "day": 96,
+  "phase": 6,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 13,
+  "weather": 1,
+  "awayOdds": 0.29425858188999554,
+  "awayOuts": 3,
+  "awayTeam": "bfd38797-8404-4b38-8b82-341da28b1f83",
+  "homeOdds": 0.7057414181100045,
+  "homeOuts": 3,
+  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
+  "outcomes": [
+      "Esme Ramsey is Partying!"
+  ],
+  "awayBalls": 4,
+  "awayBases": 4,
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 6,
+  "playCount": 115,
+  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
+  "statsheet": "8d903046-8831-4b4e-9d98-49b590e40591",
+  "atBatBalls": 3,
+  "awayBatter": null,
+  "homeBatter": "9f85676a-7411-444a-8ae2-c7f8f73c285c",
+  "lastUpdate": "Esme Ramsey is Partying!",
+  "tournament": -1,
+  "awayPitcher": "5b5bcc6c-d011-490f-b084-6fdc2c52f958",
+  "awayStrikes": 3,
+  "baseRunners": [
+      "c31d874c-1b4d-40f2-a1b3-42542e934047",
+      "a7d8196a-ca6b-4dab-a9d7-c27f3e86cc21",
+      "611d18e0-b972-4cdd-afc2-793c56bfe5a9"
+  ],
+  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "scoreLedger": "",
+  "scoreUpdate": "",
+  "seriesIndex": 1,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 2,
+  "awayTeamName": "Charleston Shoe Thieves",
+  "gameComplete": false,
+  "homeTeamName": "Canada Moist Talkers",
+  "isPostseason": false,
+  "isTitleMatch": false,
+  "seriesLength": 3,
+  "awayBatterMod": "",
+  "awayTeamColor": "#ffce0a",
+  "awayTeamEmoji": "0x1F45F",
+  "basesOccupied": [
+      2,
+      1,
+      0
+  ],
+  "homeBatterMod": "",
+  "homeTeamColor": "#f5feff",
+  "homeTeamEmoji": "0x1F5E3",
+  "awayBatterName": "",
+  "awayPitcherMod": "",
+  "baseRunnerMods": [
+      "",
+      "",
+      ""
+  ],
+  "gameStartPhase": 12,
+  "halfInningOuts": 2,
+  "homeBatterName": "Lachlan Shelton",
+  "homePitcherMod": "",
+  "newInningPhase": -1,
+  "topInningScore": 0,
+  "awayPitcherName": "Simba Davis",
+  "baseRunnerNames": [
+      "Cedric Spliff",
+      "Commissioner Vapor",
+      "Alston Cerveza"
+  ],
+  "baserunnerCount": 3,
+  "halfInningScore": 0,
+  "homePitcherName": "Greer Lott",
+  "awayTeamNickname": "Shoe Thieves",
+  "homeTeamNickname": "Moist Talkers",
+  "secretBaserunner": null,
+  "bottomInningScore": 0,
+  "awayTeamBatterCount": 7,
+  "homeTeamBatterCount": 17,
+  "awayTeamSecondaryColor": "#ffce0a",
+  "homeTeamSecondaryColor": "#f5feff"
+};
+
+const batter = {
+  "id": "b2cafd20-a799-48f6-abd7-c99bd79a6bd1",
+  "day": 116,
+  "phase": 6,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 0,
+  "season": 13,
+  "weather": 18,
+  "awayOdds": 0.44891216386293564,
+  "awayOuts": 3,
+  "awayTeam": "878c1bf6-0d21-4659-bfee-916c8314d69c",
+  "homeOdds": 0.5510878361370642,
+  "homeOuts": 3,
+  "homeTeam": "eb67ae5e-c4bf-46ca-bbbc-425cd34182ff",
+  "outcomes": [],
+  "awayBalls": 4,
+  "awayBases": 4,
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 0,
+  "playCount": 11,
+  "stadiumId": "f71fd587-d20c-461c-88a4-1634a2df3f43",
+  "statsheet": "a1d470e8-8f19-4186-a64b-73f0195a896f",
+  "atBatBalls": 0,
+  "awayBatter": "63df8701-1871-4987-87d7-b55d4f1df2e9",
+  "homeBatter": null,
+  "lastUpdate": "M---w-ll Mason batting for the Tacos.",
+  "tournament": -1,
+  "awayPitcher": "d81ce662-07b6-4a73-baa4-acbbb41f9dc5",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "e6502bc7-5b76-4939-9fb8-132057390b30",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "scoreLedger": "",
+  "scoreUpdate": "",
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "LA Unlimited Tacos",
+  "gameComplete": false,
+  "homeTeamName": "Canada Moist Talkers",
+  "isPostseason": true,
+  "isTitleMatch": false,
+  "seriesLength": 3,
+  "awayBatterMod": "",
+  "awayTeamColor": "#64376e",
+  "awayTeamEmoji": "0x1F32E",
+  "basesOccupied": [],
+  "homeBatterMod": "",
+  "homeTeamColor": "#f5feff",
+  "homeTeamEmoji": "0x1F5E3",
+  "awayBatterName": "M---w-ll Mason",
+  "awayPitcherMod": "",
+  "baseRunnerMods": [],
+  "gameStartPhase": 11,
+  "halfInningOuts": 1,
+  "homeBatterName": "",
+  "homePitcherMod": "",
+  "newInningPhase": -1,
+  "topInningScore": 0,
+  "awayPitcherName": "Yummy Elliott",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Greer Lott",
+  "awayTeamNickname": "Tacos",
+  "homeTeamNickname": "Moist Talkers",
+  "secretBaserunner": null,
+  "bottomInningScore": 0,
+  "awayTeamBatterCount": 1,
+  "homeTeamBatterCount": -1,
+  "awayTeamSecondaryColor": "#b063c1",
+  "homeTeamSecondaryColor": "#f5feff"
+};
+
+
+module.exports = {
+  maximumBlaseball,
+  batter,
+  batterAlready: 'grandSlam', // from hits
+};
+
