+++ test/data/outs.js
@@ -1,0 +1,314 @@
+const flyout = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 0,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Fish Summer hit a flyout to Penelope Mathews.",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "",
+  "halfInningOuts": 1,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": 0,
+  "homeTeamBatterCount": -1
+};
+
+const groundOut = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 2,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 0,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Spears Taylor hit a ground out to Thomas Dracaena.",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "",
+  "halfInningOuts": 0,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": 2,
+  "homeTeamBatterCount": -1
+};
+
+const strikeOut = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": null,
+  "homeBatter": null,
+  "lastUpdate": "Mclaughlin Scorpler strikes out looking.",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "",
+  "halfInningOuts": 1,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [],
+  "baserunnerCount": 0,
+  "halfInningScore": 0,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": 3,
+  "homeTeamBatterCount": 3
+};
+
+const sacrificeScore = {
+  _id: 'cbc65532-727b-4ebd-925c-d9f46d4037a0',
+  day: 57,
+  phase: 3,
+  rules: '4ae9d46a-5408-460a-84fb-cbd8d03fff6c',
+  shame: false,
+  inning: 7,
+  season: 2,
+  weather: 11,
+  awayOdds: 0.6395797544362725,
+  awayTeam: 'b72f3061-f573-40d7-832a-5ad475bd7909',
+  homeOdds: 0.3604202455637274,
+  homeTeam: '3f8bbb15-61c0-4e3f-8e4a-907a5fb1565e',
+  outcomes: [],
+  awayScore: 3,
+  finalized: false,
+  gameStart: true,
+  homeScore: 1,
+  statsheet: '16ee8bd7-3bd6-4257-b398-7e8e17a166f3',
+  atBatBalls: 0,
+  awayBatter: '',
+  homeBatter: 'f7715b05-ee69-43e5-a0e5-8e3d34270c82',
+  lastUpdate: 'Helga Burton  scores on the sacrifice.',
+  awayPitcher: '73265ee3-bb35-40d1-b696-1f241a6f5966',
+  awayStrikes: 4,
+  baseRunners: [ '58c9e294-bd49-457c-883f-fb3162fc668e' ],
+  homePitcher: 'dfd5ccbb-90ed-4bfe-83e0-dae9cc763f10',
+  homeStrikes: 3,
+  seriesIndex: 1,
+  terminology: 'b67e9bbb-1495-4e1b-b517-f1444b0a6c8b',
+  topOfInning: true,
+  atBatStrikes: 0,
+  awayTeamName: 'San Francisco Lovers',
+  gameComplete: false,
+  homeTeamName: 'Boston Flowers',
+  isPostseason: false,
+  seriesLength: 3,
+  awayTeamColor: '#780018',
+  awayTeamEmoji: '0x1F48B',
+  basesOccupied: [ 1 ],
+  homeTeamColor: '#f7d1ff',
+  homeTeamEmoji: '0x1F339',
+  awayBatterName: '',
+  halfInningOuts: 2,
+  homeBatterName: 'Caligula Lotus',
+  awayPitcherName: 'Parker Meng',
+  baserunnerCount: 1,
+  halfInningScore: 1,
+  homePitcherName: 'Owen Picklestein',
+  awayTeamNickname: 'Lovers',
+  homeTeamNickname: 'Flowers',
+  awayTeamBatterCount: 28,
+  homeTeamBatterCount: 24
+};
+
+const sacrificeAdvance = {
+  id: '8ac77e5c-fa21-4403-9b67-855d6a469e92',
+  day: 81,
+  phase: 6,
+  rules: '4ae9d46a-5408-460a-84fb-cbd8d03fff6c',
+  shame: false,
+  inning: 7,
+  season: 13,
+  weather: 18,
+  awayOdds: 0.6417733693106469,
+  awayOuts: 3,
+  awayTeam: '747b8e4a-7e50-4638-a973-ea7950a3e739',
+  homeOdds: 0.35822663068935306,
+  homeOuts: 3,
+  homeTeam: '57ec08cc-0411-4643-b304-0e80dbc15ac7',
+  outcomes: [],
+  awayBalls: 4,
+  awayBases: 4,
+  awayScore: 4,
+  finalized: false,
+  gameStart: true,
+  homeBalls: 4,
+  homeBases: 4,
+  homeScore: 2,
+  playCount: 256,
+  stadiumId: '2841dc8c-b5a2-4b3a-a717-9e646e8e29fc',
+  statsheet: '4b567f53-fc67-4f6c-9e11-028357d6d023',
+  atBatBalls: 0,
+  awayBatter: null,
+  homeBatter: null,
+  lastUpdate: 'Paula Mason hit a ground out to Ronan Combs.\n' +
+    'Randy Castillo advances on the sacrifice.',
+  tournament: -1,
+  awayPitcher: '2720559e-9173-4042-aaa0-d3852b72ab2e',
+  awayStrikes: 3,
+  baseRunners: [ 'efafe75e-2f00-4418-914c-9b6675d39264' ],
+  homePitcher: '16aff709-e855-47c8-8818-b9ba66e90fe8',
+  homeStrikes: 3,
+  repeatCount: 0,
+  scoreLedger: '',
+  scoreUpdate: '1 Run scored!',
+  seriesIndex: 1,
+  terminology: 'b67e9bbb-1495-4e1b-b517-f1444b0a6c8b',
+  topOfInning: true,
+  atBatStrikes: 0,
+  awayTeamName: 'Hades Tigers',
+  gameComplete: false,
+  homeTeamName: 'Mexico City Wild Wings',
+  isPostseason: false,
+  isTitleMatch: false,
+  seriesLength: 3,
+  awayBatterMod: '',
+  awayTeamColor: '#5c1c1c',
+  awayTeamEmoji: '0x1F405',
+  basesOccupied: [ 1 ],
+  homeBatterMod: '',
+  homeTeamColor: '#d15700',
+  homeTeamEmoji: '0x1F357',
+  awayBatterName: '',
+  awayPitcherMod: '',
+  baseRunnerMods: [ '' ],
+  gameStartPhase: 18,
+  halfInningOuts: 1,
+  homeBatterName: '',
+  homePitcherMod: '',
+  newInningPhase: -1,
+  topInningScore: 1,
+  awayPitcherName: 'Hiroto Wilcox',
+  baseRunnerNames: [ 'Aldon Cashmoney' ],
+  baserunnerCount: 1,
+  halfInningScore: 1,
+  homePitcherName: 'Mullen Peterson',
+  awayTeamNickname: 'Tigers',
+  homeTeamNickname: 'Wild Wings',
+  secretBaserunner: null,
+  bottomInningScore: 0,
+  awayTeamBatterCount: 33,
+  homeTeamBatterCount: 36,
+  awayTeamSecondaryColor: '#e83622',
+  homeTeamSecondaryColor: '#ee6300'
+};
+
+module.exports = {
+  flyout,
+  groundOut,
+  strikeOut,
+  lastOut: groundOut,
+  sacrificeScore,
+  sacrificeAdvance,
+};
+
