+++ test/data/steals.js
@@ -1,0 +1,333 @@
+const stealHome = {
+  "_id": "d8f95c60-a1c6-40f6-9d9f-25511a0f2c68",
+  "day": 102,
+  "phase": 3,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 8,
+  "season": 1,
+  "weather": 7,
+  "awayOdds": 0.5413947119936758,
+  "awayTeam": "bfd38797-8404-4b38-8b82-341da28b1f83",
+  "homeOdds": 0.4586052880063242,
+  "homeTeam": "b024e975-1c4a-4575-8936-a3754a08806a",
+  "outcomes": [],
+  "awayScore": 11,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 5,
+  "statsheet": "f69a6db4-544a-4d26-ba1e-13fd98c7249a",
+  "atBatBalls": 2,
+  "awayBatter": "18798b8f-6391-4cb2-8a5f-6fb540d646d5",
+  "homeBatter": "f73009c5-2ede-4dc4-b96d-84ba93c8a429",
+  "lastUpdate": "Marco Stink steals home!",
+  "awayPitcher": "36786f44-9066-4028-98d9-4fa84465ab9e",
+  "awayStrikes": 3,
+  "baseRunners": [
+      "17397256-c28c-4cad-85f2-a21768c66e67",
+      "c83f0fe0-44d1-4342-81e8-944bb38f8e23"
+  ],
+  "homePitcher": "740d5fef-d59f-4dac-9a75-739ec07f91cf",
+  "homeStrikes": 3,
+  "seriesIndex": 4,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": false,
+  "atBatStrikes": 1,
+  "awayTeamName": "Charleston Shoe Thieves",
+  "gameComplete": false,
+  "homeTeamName": "Dallas Steaks",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#ffce0a",
+  "awayTeamEmoji": "0x1F45F",
+  "basesOccupied": [
+      1,
+      0
+  ],
+  "homeTeamColor": "#8c8d8f",
+  "homeTeamEmoji": "0x1F969",
+  "awayBatterName": "Morrow Doyle",
+  "halfInningOuts": 2,
+  "homeBatterName": "Thomas Kirby",
+  "awayPitcherName": "Beasley Gloom",
+  "baserunnerCount": 2,
+  "halfInningScore": 1,
+  "homePitcherName": "Conner Haley",
+  "awayTeamNickname": "Shoe Thieves",
+  "homeTeamNickname": "Steaks",
+  "awayTeamBatterCount": 45,
+  "homeTeamBatterCount": 38
+};
+
+const caughtStealing = {
+  "_id": "72418a6e-b2c0-44dc-bc9f-e045e049411b",
+  "day": 104,
+  "phase": 3,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 6,
+  "season": 1,
+  "weather": 7,
+  "awayOdds": 0.4803003352967402,
+  "awayTeam": "23e4cbc1-e9cd-47fa-a35b-bfa06f726cb7",
+  "homeOdds": 0.5196996647032597,
+  "homeTeam": "a37f9158-7f82-46bc-908c-c9e2dda7c33b",
+  "outcomes": [],
+  "awayScore": 6,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 7,
+  "statsheet": "a297f78b-b9f5-43c4-b989-f3c9df0f415a",
+  "atBatBalls": 0,
+  "awayBatter": "15ae64cd-f698-4b00-9d61-c9fffd037ae2",
+  "homeBatter": "",
+  "lastUpdate": "Elvis Figueroa gets caught stealing third base.",
+  "awayPitcher": "814bae61-071a-449b-981e-e7afc839d6d6",
+  "awayStrikes": 3,
+  "baseRunners": [
+      "667cb445-c288-4e62-b603-27291c1e475d"
+  ],
+  "homePitcher": "f4a5d734-0ade-4410-abb6-c0cd5a7a1c26",
+  "homeStrikes": 3,
+  "seriesIndex": 1,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Philly Pies",
+  "gameComplete": false,
+  "homeTeamName": "Breckenridge Jazz Hands",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#399d8f",
+  "awayTeamEmoji": "0x1F967",
+  "basesOccupied": [
+      0
+  ],
+  "homeTeamColor": "#6388ad",
+  "homeTeamEmoji": "0x1F450",
+  "awayBatterName": "Mickey Woods",
+  "halfInningOuts": 2,
+  "homeBatterName": "",
+  "awayPitcherName": "Ruslan Greatness",
+  "baserunnerCount": 1,
+  "halfInningScore": 3,
+  "homePitcherName": "Agan Harrison",
+  "awayTeamNickname": "Pies",
+  "homeTeamNickname": "Jazz Hands",
+  "awayTeamBatterCount": 32,
+  "homeTeamBatterCount": 26
+};
+const steal = {
+  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
+  "day": 119,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 10,
+  "weather": 14,
+  "awayOdds": 0.38125626427142956,
+  "awayOuts": 3,
+  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
+  "homeOdds": 0.6187437357285706,
+  "homeOuts": 3,
+  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
+  "outcomes": [],
+  "awayBalls": 3,
+  "awayBases": 4,
+  "awayScore": 1,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 0,
+  "playCount": 43,
+  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
+  "atBatBalls": 0,
+  "awayBatter": "190a0f31-d686-4ac4-a7f3-cfc87b72c145",
+  "homeBatter": null,
+  "lastUpdate": "Sutton Bishop steals fourth base!",
+  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
+  "awayStrikes": 4,
+  "baseRunners": [
+      "4f69e8c2-b2a1-4e98-996a-ccf35ac844c5"
+  ],
+  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "seriesIndex": 5,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hellmouth Sunbeams",
+  "gameComplete": false,
+  "homeTeamName": "Seattle Garages",
+  "isPostseason": true,
+  "seriesLength": 3,
+  "awayTeamColor": "#fffbab",
+  "awayTeamEmoji": "0x1F31E",
+  "basesOccupied": [
+      0
+  ],
+  "homeTeamColor": "#2b4075",
+  "homeTeamEmoji": "0x1F3B8",
+  "awayBatterName": "Nerd Pacheco",
+  "halfInningOuts": 1,
+  "homeBatterName": "",
+  "awayPitcherName": "Lars Taylor",
+  "baseRunnerNames": [
+      "Igneus Delacruz"
+  ],
+  "baserunnerCount": 1,
+  "halfInningScore": 1,
+  "homePitcherName": "Arturo Huerta",
+  "awayTeamNickname": "Sunbeams",
+  "homeTeamNickname": "Garages",
+  "awayTeamBatterCount": 6,
+  "homeTeamBatterCount": 3,
+  "awayTeamSecondaryColor": "#fffbab",
+  "homeTeamSecondaryColor": "#5a83ea"
+};
+
+const stealScore = {
+  "id": "47bcac42-f651-4fc9-9f93-5567a7b10daf",
+  "day": 119,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 10,
+  "weather": 14,
+  "awayOdds": 0.38125626427142956,
+  "awayOuts": 3,
+  "awayTeam": "f02aeae2-5e6a-4098-9842-02d2273f25c7",
+  "homeOdds": 0.6187437357285706,
+  "homeOuts": 3,
+  "homeTeam": "105bc3ff-1320-4e37-8ef0-8d595cb95dd0",
+  "outcomes": [],
+  "awayBalls": 3,
+  "awayBases": 4,
+  "awayScore": 1,
+  "finalized": false,
+  "gameStart": true,
+  "homeBalls": 4,
+  "homeBases": 4,
+  "homeScore": 0,
+  "playCount": 43,
+  "statsheet": "aef3a6d6-768b-434d-88ce-8aecda7ee503",
+  "atBatBalls": 0,
+  "awayBatter": "190a0f31-d686-4ac4-a7f3-cfc87b72c145",
+  "homeBatter": null,
+  "lastUpdate": "Sutton Bishop steals fourth base!",
+  "awayPitcher": "20fd71e7-4fa0-4132-9f47-06a314ed539a",
+  "awayStrikes": 4,
+  "baseRunners": [
+      "4f69e8c2-b2a1-4e98-996a-ccf35ac844c5"
+  ],
+  "homePitcher": "6f9de777-e812-4c84-915c-ef283c9f0cde",
+  "homeStrikes": 3,
+  "repeatCount": 0,
+  "seriesIndex": 5,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hellmouth Sunbeams",
+  "gameComplete": false,
+  "homeTeamName": "Seattle Garages",
+  "isPostseason": true,
+  "seriesLength": 3,
+  "awayTeamColor": "#fffbab",
+  "awayTeamEmoji": "0x1F31E",
+  "basesOccupied": [
+      0
+  ],
+  "homeTeamColor": "#2b4075",
+  "homeTeamEmoji": "0x1F3B8",
+  "awayBatterName": "Nerd Pacheco",
+  "halfInningOuts": 1,
+  "homeBatterName": "",
+  "awayPitcherName": "Lars Taylor",
+  "baseRunnerNames": [
+      "Igneus Delacruz"
+  ],
+  "baserunnerCount": 1,
+  "halfInningScore": 1,
+  "homePitcherName": "Arturo Huerta",
+  "awayTeamNickname": "Sunbeams",
+  "homeTeamNickname": "Garages",
+  "awayTeamBatterCount": 6,
+  "homeTeamBatterCount": 3,
+  "awayTeamSecondaryColor": "#fffbab",
+  "homeTeamSecondaryColor": "#5a83ea"
+};
+
+const stealSecond = {
+  "id": "628a2ddb-f608-411b-8d2e-2768cd36d58b",
+  "day": 111,
+  "phase": 5,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 3,
+  "weather": 12,
+  "awayOdds": 0.5588613238139439,
+  "awayTeam": "747b8e4a-7e50-4638-a973-ea7950a3e739",
+  "homeOdds": 0.441138676186056,
+  "homeTeam": "36569151-a2fb-43c1-9df7-2df512424c82",
+  "outcomes": [],
+  "awayScore": 0,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "5f28188f-51c7-4a1f-bd3e-3ebe1912dfdf",
+  "atBatBalls": 0,
+  "awayBatter": "32c9bce6-6e52-40fa-9f64-3629b3d026a8",
+  "homeBatter": null,
+  "lastUpdate": "Moody Cookbook steals second base!",
+  "awayPitcher": "afc90398-b891-4cdf-9dea-af8a3a79d793",
+  "awayStrikes": 3,
+  "baseRunners": [
+      "70ccff1e-6b53-40e2-8844-0a28621cb33e"
+  ],
+  "homePitcher": "81d7d022-19d6-427d-aafc-031fcb79b29e",
+  "homeStrikes": 3,
+  "seriesIndex": 3,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Hades Tigers",
+  "gameComplete": false,
+  "homeTeamName": "New York Millennials",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#5c1c1c",
+  "awayTeamEmoji": "0x1F405",
+  "basesOccupied": [
+      1
+  ],
+  "homeTeamColor": "#ffd4d8",
+  "homeTeamEmoji": "0x1F4F1",
+  "awayBatterName": "Ren Morin",
+  "halfInningOuts": 2,
+  "homeBatterName": "",
+  "awayPitcherName": "Yazmin Mason",
+  "baseRunnerNames": [
+      "Moody Cookbook"
+  ],
+  "baserunnerCount": 1,
+  "halfInningScore": 0,
+  "homePitcherName": "Patty Fox",
+  "awayTeamNickname": "Tigers",
+  "homeTeamNickname": "Millennials",
+  "awayTeamBatterCount": 6,
+  "homeTeamBatterCount": 3
+};
+module.exports = {
+  steal,
+  stealScore: steal,
+  stealScore,
+  stealSecond,
+  stealHome,
+  caughtStealing,
+};
+
