+++ test/data/walks.js
@@ -1,0 +1,151 @@
+const walk = {
+  "_id": "97d88b9e-406d-4f31-a18f-2a3b903edc03",
+  "day": 110,
+  "phase": 3,
+  "rules": "4ae9d46a-5408-460a-84fb-cbd8d03fff6c",
+  "shame": false,
+  "inning": 1,
+  "season": 1,
+  "weather": 7,
+  "awayOdds": 0.6557171270683066,
+  "awayTeam": "23e4cbc1-e9cd-47fa-a35b-bfa06f726cb7",
+  "homeOdds": 0.34428287293169346,
+  "homeTeam": "bfd38797-8404-4b38-8b82-341da28b1f83",
+  "outcomes": [],
+  "awayScore": 1,
+  "finalized": false,
+  "gameStart": true,
+  "homeScore": 0,
+  "statsheet": "9a42badd-8646-4e90-9d33-3e5accc3403f",
+  "atBatBalls": 0,
+  "awayBatter": "",
+  "homeBatter": "248ccf3d-d5f6-4b69-83d9-40230ca909cd",
+  "lastUpdate": "Jessica Telephone draws a walk.",
+  "awayPitcher": "1732e623-ffc2-40f0-87ba-fdcf97131f1f",
+  "awayStrikes": 3,
+  "baseRunners": [
+      "083d09d4-7ed3-4100-b021-8fbe30dd43e8"
+  ],
+  "homePitcher": "2ae8cbfc-2155-4647-9996-3f2591091baf",
+  "homeStrikes": 3,
+  "seriesIndex": 4,
+  "terminology": "b67e9bbb-1495-4e1b-b517-f1444b0a6c8b",
+  "topOfInning": true,
+  "atBatStrikes": 0,
+  "awayTeamName": "Philly Pies",
+  "gameComplete": false,
+  "homeTeamName": "Charleston Shoe Thieves",
+  "isPostseason": true,
+  "seriesLength": 5,
+  "awayTeamColor": "#399d8f",
+  "awayTeamEmoji": "0x1F967",
+  "basesOccupied": [
+      0
+  ],
+  "homeTeamColor": "#ffce0a",
+  "homeTeamEmoji": "0x1F45F",
+  "awayBatterName": "",
+  "halfInningOuts": 1,
+  "homeBatterName": "Antonio Wallace",
+  "awayPitcherName": "Betsy Trombone",
+  "baserunnerCount": 1,
+  "halfInningScore": 0,
+  "homePitcherName": "Forrest Bookbaby",
+  "awayTeamNickname": "Pies",
+  "homeTeamNickname": "Shoe Thieves",
+  "awayTeamBatterCount": 6,
+  "homeTeamBatterCount": 3
+};
+
+const walkScores = {
+  id: 'b6f5af1a-79e8-4f05-a6c7-9b55c42124ab',
+  day: 95,
+  phase: 6,
+  rules: '4ae9d46a-5408-460a-84fb-cbd8d03fff6c',
+  shame: false,
+  inning: 0,
+  season: 13,
+  weather: 13,
+  awayOdds: 0.5538712361406611,
+  awayOuts: 3,
+  awayTeam: 'eb67ae5e-c4bf-46ca-bbbc-425cd34182ff',
+  homeOdds: 0.4461287638593388,
+  homeOuts: 3,
+  homeTeam: 'adc5b394-8f76-416d-9ce9-813706877b84',
+  outcomes: [ 'Rodriguez Internet returned from Elsewhere!' ],
+  awayBalls: 4,
+  awayBases: 4,
+  awayScore: 1,
+  finalized: false,
+  gameStart: true,
+  homeBalls: 4,
+  homeBases: 4,
+  homeScore: 6,
+  playCount: 83,
+  stadiumId: '21ce1233-b2d6-4bd2-8d03-68a8915afdc5',
+  statsheet: '8778af73-a34a-4f67-870f-9701f3e1d0d8',
+  atBatBalls: 0,
+  awayBatter: null,
+  homeBatter: null,
+  lastUpdate: 'Hewitt Best draws a walk.\n' +
+    'Marquez Clark scores!\n' +
+    'Hewitt Best used their Free Refill.\n' +
+    'Hewitt Best Refills the In!',
+  tournament: -1,
+  awayPitcher: '60026a9d-fc9a-4f5a-94fd-2225398fa3da',
+  awayStrikes: 3,
+  baseRunners: [
+    '53e701c7-e3c8-4e18-ba05-9b41b4b64cda',
+    'd6c69d2d-9344-4b19-85a4-6cfcbaead5d2',
+    '88cd6efa-dbf2-4309-aabe-ec1d6f21f98a'
+  ],
+  homePitcher: '138fccc3-e66f-4b07-8327-d4b6f372f654',
+  homeStrikes: 3,
+  repeatCount: 0,
+  scoreLedger: '',
+  scoreUpdate: '1 Run scored!',
+  seriesIndex: 3,
+  terminology: 'b67e9bbb-1495-4e1b-b517-f1444b0a6c8b',
+  topOfInning: false,
+  atBatStrikes: 0,
+  awayTeamName: 'Canada Moist Talkers',
+  gameComplete: false,
+  homeTeamName: 'Kansas City Breath Mints',
+  isPostseason: false,
+  isTitleMatch: false,
+  seriesLength: 3,
+  awayBatterMod: '',
+  awayTeamColor: '#f5feff',
+  awayTeamEmoji: '0x1F5E3',
+  basesOccupied: [ 2, 1, 0 ],
+  homeBatterMod: '',
+  homeTeamColor: '#178f55',
+  homeTeamEmoji: '0x1F36C',
+  awayBatterName: '',
+  awayPitcherMod: '',
+  baseRunnerMods: [ '', '', '' ],
+  gameStartPhase: 11,
+  halfInningOuts: 1,
+  homeBatterName: '',
+  homePitcherMod: '',
+  newInningPhase: -1,
+  topInningScore: 1,
+  awayPitcherName: 'Bright Zimmerman',
+  baseRunnerNames: [ 'Marquez Clark', 'Joshua Watson', 'Hewitt Best' ],
+  baserunnerCount: 3,
+  halfInningScore: 6,
+  homePitcherName: 'Oscar Vaughan',
+  awayTeamNickname: 'Moist Talkers',
+  homeTeamNickname: 'Breath Mints',
+  secretBaserunner: null,
+  bottomInningScore: 6,
+  awayTeamBatterCount: 4,
+  homeTeamBatterCount: 8,
+  awayTeamSecondaryColor: '#f5feff',
+  homeTeamSecondaryColor: '#178f55'
+};
+
+module.exports = {
+  walk,
+  walkScores,
+};
